% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{add_jumpstart_tags}
\alias{add_jumpstart_tags}
\title{Add custom tags to JumpStart models, return the updated tags.}
\usage{
add_jumpstart_tags(
  tags = NULL,
  inference_model_uri = NULL,
  inference_script_uri = NULL,
  training_model_uri = NULL,
  training_script_uri = NULL
)
}
\arguments{
\item{tags}{(Optional[List[Dict[str,str]]): Current tags for JumpStart inference
or training job. (Default: None).}

\item{inference_model_uri}{(Optional[str]): S3 URI for inference model artifact.
(Default: None).}

\item{inference_script_uri}{(Optional[str]): S3 URI for inference script tarball.
(Default: None).}

\item{training_model_uri}{(Optional[str]): S3 URI for training model artifact.
(Default: None).}

\item{training_script_uri}{(Optional[str]): S3 URI for training script tarball.
(Default: None).}
}
\description{
No-op if this is not a JumpStart model related resource.
}
