% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{add_single_jumpstart_tag}
\alias{add_single_jumpstart_tag}
\title{Adds ``tag_key`` to ``curr_tags`` if ``uri`` corresponds to a JumpStart model.}
\usage{
add_single_jumpstart_tag(uri, tag_key, curr_tags)
}
\arguments{
\item{uri}{(str): URI which may correspond to a JumpStart model.}

\item{tag_key}{(enums.JumpStartTag): Custom tag to apply to current tags if the URI
corresponds to a JumpStart model.}

\item{curr_tags}{(Optional[List]): Current tags associated with ``Estimator`` or ``Model``.}
}
\description{
Adds ``tag_key`` to ``curr_tags`` if ``uri`` corresponds to a JumpStart model.
}
