% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiutils_base_types.R
\name{ApiObject}
\alias{ApiObject}
\title{A Python class representation of a boto API object.}
\description{
Converts boto dicts of 'UpperCamelCase' names to dicts into/from a Python object with standard
             python members. Clients invoke to_boto on an instance of ApiObject to transform the ApiObject
             into a boto representation. Clients invoke from_boto on a sub-class of ApiObject to
             instantiate an instance of that class from a boto representation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApiObject$new()}}
\item \href{#method-from_paws}{\code{ApiObject$from_paws()}}
\item \href{#method-to_paws}{\code{ApiObject$to_paws()}}
\item \href{#method-format}{\code{ApiObject$format()}}
\item \href{#method-clone}{\code{ApiObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ApiObject class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiObject$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_paws"></a>}}
\if{latex}{\out{\hypertarget{method-from_paws}{}}}
\subsection{Method \code{from_paws()}}{
Construct an instance of this ApiObject from a boto response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiObject$from_paws(paws_list, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paws_list}}{(list): A dictionary of a paws response.}

\item{\code{...}}{: Arbitrary keyword arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_paws"></a>}}
\if{latex}{\out{\hypertarget{method-to_paws}{}}}
\subsection{Method \code{to_paws()}}{
Convert an object to a paws representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiObject$to_paws(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{(dict): The object to convert to paws.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Return a string representation of this ApiObject.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiObject$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
