% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{BatchStrategy}
\alias{BatchStrategy}
\title{BatchStrategy class}
\description{
BatchStrategy class

BatchStrategy class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchStrategy-new}{\code{BatchStrategy$new()}}
\item \href{#method-BatchStrategy-pad}{\code{BatchStrategy$pad()}}
\item \href{#method-BatchStrategy-clone}{\code{BatchStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchStrategy-new}{}}}
\subsection{Method \code{new()}}{
Create a Batch Strategy Instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchStrategy$new(splitter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{splitter}}{(sagemaker.local.data.Splitter): A Splitter to pre-process
the data before batching.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchStrategy-pad"></a>}}
\if{latex}{\out{\hypertarget{method-BatchStrategy-pad}{}}}
\subsection{Method \code{pad()}}{
Group together as many records as possible to fit in the specified size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchStrategy$pad(file, size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): file path to read the records from.}

\item{\code{size}}{(int): maximum size in MB that each group of records will be
fitted to. passing 0 means unlimited size.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generator of records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
