% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\docType{data}
\name{BooleanValues}
\alias{BooleanValues}
\title{Enum class for boolean values.}
\format{
An object of class \code{BooleanValues} (inherits from \code{Enum}, \code{environment}) of length 4.
}
\usage{
BooleanValues
}
\description{
This is a status value that an ``Operand`` can resolve to.
}
\keyword{datasets}
