% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_dict}
\alias{build_dict}
\title{Return a dict of key and value pair if value is not None, otherwise return an empty dict.}
\usage{
build_dict(key, value = NULL)
}
\arguments{
\item{key}{(str): input key}

\item{value}{(str): input value}
}
\value{
dict: dict of key and value or an empty dict.
}
\description{
Return a dict of key and value pair if value is not None, otherwise return an empty dict.
}
\seealso{
Other sagemaker_utils: 
\code{\link{.aws_partition}()},
\code{\link{.download_files_under_prefix}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{common_variables}},
\code{\link{create_tar_file}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_config_value}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
