% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{cls_help}
\alias{cls_help}
\title{Helper function to return help documentation for sagemaker R6 classes.}
\usage{
cls_help(cls)
}
\arguments{
\item{cls}{(R6::R6Class): R6 class}
}
\description{
Helper function to return help documentation for sagemaker R6 classes.
}
\seealso{
Other r_utils: 
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{pkg_method}()},
\code{\link{rsplit}()}
}
\concept{r_utils}
