% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{common_varaibles}
\alias{common_varaibles}
\alias{ECR_URI_PATTERN}
\alias{MAX_BUCKET_PATHS_COUNT}
\alias{S3_PREFIX}
\alias{HTTP_PREFIX}
\alias{HTTPS_PREFIX}
\alias{DEFAULT_SLEEP_TIME_SECONDS}
\title{Core varaibles}
\format{
An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
ECR_URI_PATTERN

MAX_BUCKET_PATHS_COUNT

S3_PREFIX

HTTP_PREFIX

HTTPS_PREFIX

DEFAULT_SLEEP_TIME_SECONDS
}
\description{
Core varaibles
}
\keyword{datasets}
\keyword{internal}
