% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{common_variables}
\alias{common_variables}
\alias{ECR_URI_PATTERN}
\alias{MAX_BUCKET_PATHS_COUNT}
\alias{S3_PREFIX}
\alias{HTTP_PREFIX}
\alias{HTTPS_PREFIX}
\alias{DEFAULT_SLEEP_TIME_SECONDS}
\title{Core variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
ECR_URI_PATTERN

MAX_BUCKET_PATHS_COUNT

S3_PREFIX

HTTP_PREFIX

HTTPS_PREFIX

DEFAULT_SLEEP_TIME_SECONDS
}
\description{
Core variables
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
\keyword{datasets}
\keyword{internal}
