% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Constant}
\alias{Constant}
\title{Constant operator class for filtering JumpStart content.}
\description{
Constant operator class for filtering JumpStart content.

Constant operator class for filtering JumpStart content.
}
\section{Super classes}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{\link[sagemaker.core:Operator]{sagemaker.core::Operator}} -> \code{Constant}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Constant-new}{\code{Constant$new()}}
\item \href{#method-Constant-eval}{\code{Constant$eval()}}
\item \href{#method-Constant-clone}{\code{Constant$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format"><a href='../../sagemaker.core/html/Operand.html#method-Operand-format'><code>sagemaker.core::Operand$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand"><a href='../../sagemaker.core/html/Operand.html#method-Operand-validate_operand'><code>sagemaker.core::Operand$validate_operand()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Constant-new"></a>}}
\if{latex}{\out{\hypertarget{method-Constant-new}{}}}
\subsection{Method \code{new()}}{
Instantiates Constant operator object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$new(constant)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{constant}}{(BooleanValues): Value of constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Constant-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Constant-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates constant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Constant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Constant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
