% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_utils.R
\name{copy_directory_structure}
\alias{copy_directory_structure}
\title{Creates intermediate directory structure for relative_path.}
\usage{
copy_directory_structure(destination_directory, relative_path)
}
\arguments{
\item{destination_directory}{(str): root of the destination directory where the
directory structure will be created.}

\item{relative_path}{(str): relative path that will be created within
destination_directory}
}
\description{
Create all the intermediate directories required for relative_path to
             exist within destination_directory. This assumes that relative_path is a
             directory located within root_dir.
}
