% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tar_file}
\alias{create_tar_file}
\title{Create a tar file containing all the source_files}
\usage{
create_tar_file(source_files, target = NULL)
}
\arguments{
\item{source_files}{(str): vector of file paths that will be contained in the tar file}

\item{target}{(str): A character vector of filepaths to be archived. Default to archive all files
into a temporary file.}
}
\value{
(str): path to created tar file
}
\description{
Create a tar file containing all the source_files
}
\seealso{
Other sagemaker_utils: 
\code{\link{.aws_partition}()},
\code{\link{.download_files_under_prefix}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_config_value}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
\keyword{internal}
