% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{CreateModelInput}
\alias{CreateModelInput}
\title{CreateModelInput}
\description{
A class containing parameters which can be used to create a SageMaker Model
             Parameters:
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instance_type}}{type or EC2 instance will be used for model deployment}

\item{\code{accelerator_type}}{elastic inference accelerator type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateModelInput$new()}}
\item \href{#method-format}{\code{CreateModelInput$format()}}
\item \href{#method-clone}{\code{CreateModelInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize CreateModelInput class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateModelInput$new(instance_type = NULL, accelerator_type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance_type}}{(str): type or EC2 instance will be used for model deployment.}

\item{\code{accelerator_type}}{(str): elastic inference accelerator type.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateModelInput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateModelInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
