% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_exceptions.R
\name{DeprecatedJumpStartModelError}
\alias{DeprecatedJumpStartModelError}
\title{Exception raised when trying to access a JumpStart model deprecated specifications.}
\description{
A deprecated specification for a JumpStart model does not mean the whole model is
             deprecated. There may be more recent specifications available for this model. For
             example, all specification before version ``2.0.0`` may be deprecated, in such a
             case, the SDK would raise this exception only when specifications ``1.*`` are
             accessed.
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{DeprecatedJumpStartModelError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeprecatedJumpStartModelError-new}{\code{DeprecatedJumpStartModelError$new()}}
\item \href{#method-DeprecatedJumpStartModelError-clone}{\code{DeprecatedJumpStartModelError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedJumpStartModelError-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedJumpStartModelError-new}{}}}
\subsection{Method \code{new()}}{
Instantiates DeprecatedJumpStartModelError exception.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedJumpStartModelError$new(
  model_id = NULL,
  version = NULL,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{(Optional[str]): model ID of vulnerable JumpStart model.
(Default: None).}

\item{\code{version}}{(Optional[str]): version of vulnerable JumpStart model.
(Default: None).}

\item{\code{message}}{(Optional[str]): error message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedJumpStartModelError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedJumpStartModelError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedJumpStartModelError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
