% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{.append_project_tags}
\alias{.append_project_tags}
\title{Appends the project tag to the list of tags, if it exists.}
\usage{
.append_project_tags(tags = NULL, working_dir = NULL)
}
\arguments{
\item{tags}{: the list of tags to append to.}

\item{working_dir}{: the working directory to start looking.}
}
\value{
A possibly extended list of tags that includes the project id
}
\description{
Appends the project tag to the list of tags, if it exists.
}
