% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.aws_partition}
\alias{.aws_partition}
\title{Given a region name (ex: "cn-north-1"), return the corresponding aws partition ("aws-cn").}
\usage{
.aws_partition(region)
}
\value{
str: partition corresponding to the region name passed in.
}
\description{
region (str): The region name for which to return the corresponding partition.
}
\seealso{
Other sagemaker_utils: 
\code{\link{.download_files_under_prefix}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{create_tar_file}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_config_value}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
\keyword{internal}
