% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.download_files_under_prefix}
\alias{.download_files_under_prefix}
\title{Download all S3 files which match the given prefix}
\usage{
.download_files_under_prefix(bucket_name, prefix, target, s3)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{prefix}{(str): S3 prefix within the bucket that will be downloaded}

\item{target}{(str): destination path where the downloaded items will be placed}

\item{s3}{(paws::s3): S3 resource}
}
\description{
Download all S3 files which match the given prefix
}
\seealso{
Other sagemaker_utils: 
\code{\link{.aws_partition}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{create_tar_file}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_config_value}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
\keyword{internal}
