% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_notebook_utils.R
\name{.generate_jumpstart_model_versions}
\alias{.generate_jumpstart_model_versions}
\title{Generate models for JumpStart, and optionally apply filters to result.}
\usage{
.generate_jumpstart_model_versions(
  filter = Constant$new(BooleanValues$`TRUE`),
  region = JUMPSTART_DEFAULT_REGION_NAME(),
  list_incomplete_models = FALSE
)
}
\arguments{
\item{filter}{(Union[Operator, str]): Optional. The filter to apply to generate models. This can be
either an ``Operator`` type filter (e.g. ``And("task == ic", "framework == pytorch")``),
or simply a string filter which will get serialized into an Identity filter.
(e.g. ``"task == ic"``). If this argument is not supplied, all models will be generated.
(Default: Constant(BooleanValues$TRUE)).}

\item{region}{(str): Optional. The AWS region from which to retrieve JumpStart metadata regarding
models. (Default: JUMPSTART_DEFAULT_REGION_NAME()).}

\item{list_incomplete_models}{(bool): Optional. If a model does not contain metadata fields
requested by the filter, and the filter cannot be resolved to a include/not include,
whether the model should be included. By default, these models are omitted from
results. (Default: False).}
}
\description{
Generate models for JumpStart, and optionally apply filters to result.
}
