% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_accessors.R
\name{.JumpStartModelsAccessor}
\alias{.JumpStartModelsAccessor}
\title{JumpStartModelsAccessor class}
\description{
Static class for storing the JumpStart models cache.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartModelsAccessor-get_model_header}{\code{.JumpStartModelsAccessor$get_model_header()}}
\item \href{#method-JumpStartModelsAccessor-get_model_specs}{\code{.JumpStartModelsAccessor$get_model_specs()}}
\item \href{#method-JumpStartModelsAccessor-set_cache_kwargs}{\code{.JumpStartModelsAccessor$set_cache_kwargs()}}
\item \href{#method-JumpStartModelsAccessor-reset_cache}{\code{.JumpStartModelsAccessor$reset_cache()}}
\item \href{#method-JumpStartModelsAccessor-get_manifest}{\code{.JumpStartModelsAccessor$get_manifest()}}
\item \href{#method-JumpStartModelsAccessor-clone}{\code{.JumpStartModelsAccessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-get_model_header"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-get_model_header}{}}}
\subsection{Method \code{get_model_header()}}{
Returns model header from JumpStart models cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$get_model_header(region, model_id, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{(str): region for which to retrieve header.}

\item{\code{model_id}}{(str): model id to retrieve.}

\item{\code{version}}{(str): semantic version to retrieve for the model id.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-get_model_specs"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-get_model_specs}{}}}
\subsection{Method \code{get_model_specs()}}{
Returns model specs from JumpStart models cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$get_model_specs(region, model_id, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{(str): region for which to retrieve header.}

\item{\code{model_id}}{(str): model id to retrieve.}

\item{\code{version}}{(str): semantic version to retrieve for the model id.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-set_cache_kwargs"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-set_cache_kwargs}{}}}
\subsection{Method \code{set_cache_kwargs()}}{
Sets cache kwargs, clears the cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$set_cache_kwargs(cache_kwargs, region = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_kwargs}}{(str): cache kwargs to validate.}

\item{\code{region}}{(str): Optional. The region to validate along with the kwargs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-reset_cache"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-reset_cache}{}}}
\subsection{Method \code{reset_cache()}}{
Resets cache, optionally allowing cache kwargs to be passed to the new cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$reset_cache(cache_kwargs = NULL, region = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_kwargs}}{(str): cache kwargs to validate.}

\item{\code{region}}{(str): The region to validate along with the kwargs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-get_manifest"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-get_manifest}{}}}
\subsection{Method \code{get_manifest()}}{
Return entire JumpStart models manifest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$get_manifest(cache_kwargs = NULL, region = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_kwargs}}{(Dict[str, Any]): Optional. Cache kwargs to use.
(Default: None).}

\item{\code{region}}{(str): Optional. The region to use for the cache.
(Default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsAccessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsAccessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.JumpStartModelsAccessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
