% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_image.R
\name{.pull_image}
\alias{.pull_image}
\title{Invokes the docker pull command for the given image.}
\usage{
.pull_image(image)
}
\arguments{
\item{image}{(str): pull docker image}
}
\description{
Invokes the docker pull command for the given image.
}
