% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{.region_supports_debugger}
\alias{.region_supports_debugger}
\title{Returns boolean indicating whether the region supports Amazon SageMaker Debugger.}
\usage{
.region_supports_debugger(region_name)
}
\arguments{
\item{region_name}{(str): Name of the region to check against.}
}
\value{
bool: Whether or not the region supports Amazon SageMaker Debugger.
}
\description{
Returns boolean indicating whether the region supports Amazon SageMaker Debugger.
}
\keyword{internal}
