% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{.region_supports_profiler}
\alias{.region_supports_profiler}
\title{Returns bool indicating whether region supports Amazon SageMaker Debugger profiling feature.}
\usage{
.region_supports_profiler(region_name)
}
\arguments{
\item{region_name}{(str): Name of the region to check against.}
}
\value{
bool: Whether or not the region supports Amazon SageMaker Debugger profiling feature.
}
\description{
Returns bool indicating whether region supports Amazon SageMaker Debugger profiling feature.
}
\keyword{internal}
