% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_accessors.R
\name{.SageMakerSettings}
\alias{.SageMakerSettings}
\title{SageMakerSettings class}
\description{
Static class for storing the SageMaker settings.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SageMakerSettings-set_sagemaker_version}{\code{.SageMakerSettings$set_sagemaker_version()}}
\item \href{#method-SageMakerSettings-get_sagemaker_version}{\code{.SageMakerSettings$get_sagemaker_version()}}
\item \href{#method-SageMakerSettings-clone}{\code{.SageMakerSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SageMakerSettings-set_sagemaker_version"></a>}}
\if{latex}{\out{\hypertarget{method-SageMakerSettings-set_sagemaker_version}{}}}
\subsection{Method \code{set_sagemaker_version()}}{
Set SageMaker version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.SageMakerSettings$set_sagemaker_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{(str): python sagemaker version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SageMakerSettings-get_sagemaker_version"></a>}}
\if{latex}{\out{\hypertarget{method-SageMakerSettings-get_sagemaker_version}{}}}
\subsection{Method \code{get_sagemaker_version()}}{
Return SageMaker version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.SageMakerSettings$get_sagemaker_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SageMakerSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SageMakerSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.SageMakerSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
