% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{.validate_payload_size}
\alias{.validate_payload_size}
\title{Check if a payload is within the size in MB threshold.}
\usage{
.validate_payload_size(payload, size)
}
\arguments{
\item{payload}{: data that will be checked}

\item{size}{(int): max size in MB}
}
\value{
bool: True if within bounds. if size=0 it will always return True
}
\description{
Raise an exception if the payload is beyond the size in MB threshold.
}
