% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{.validate_smdataparallel_args}
\alias{.validate_smdataparallel_args}
\title{Check if request is using unsupported arguments.}
\usage{
.validate_smdataparallel_args(
  instance_type,
  framework_name,
  framework_version,
  py_version,
  distribution,
  image_uri = NULL
)
}
\arguments{
\item{instance_type}{(str): A string representing the type of training instance selected. Ex: `ml.p3.16xlarge`}

\item{framework_name}{(str): A string representing the name of framework selected. Ex: `tensorflow`}

\item{framework_version}{(str): A string representing the framework version selected. Ex: `2.3.1`}

\item{py_version}{(str): A string representing the python version selected. Ex: `py3`}

\item{distribution}{(dict): A dictionary with information to enable distributed training.
(Defaults to None if distributed training is not enabled.)}
}
\description{
Validate if user specifies a supported instance type, framework version, and python
             version.
}
\keyword{internal}
