% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a Single File from S3 into a local path}
\usage{
download_file(bucket_name, path, target, sagemaker_session)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{path}{(str): file path within the bucket}

\item{target}{(str): destination directory for the downloaded file.}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}
}
\description{
Download a Single File from S3 into a local path
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
