% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file_from_url}
\alias{download_file_from_url}
\title{Download a Single File from S3 into a local path}
\usage{
download_file_from_url(url, dst, sagemaker_session)
}
\arguments{
\item{url}{(str): file path within the bucket}

\item{dst}{(str): destination directory for the downloaded file.}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}
}
\description{
Download a Single File from S3 into a local path
}
