% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_folder}
\alias{download_folder}
\title{Download a folder from S3 to a local path}
\usage{
download_folder(bucket_name, prefix, target, sagemaker_session)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{prefix}{(str): S3 prefix within the bucket that will be downloaded. Can
be a single file.}

\item{target}{(str): destination path where the downloaded items will be placed}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}
}
\description{
Download a folder from S3 to a local path
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{download_file}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
