% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_cache.R
\name{Element}
\alias{Element}
\title{Class describes the values in the cache.}
\description{
This object stores the value itself as well as a timestamp so that this
             element can be invalidated if it becomes too old.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Element-new}{\code{Element$new()}}
\item \href{#method-Element-clone}{\code{Element$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Element-new"></a>}}
\if{latex}{\out{\hypertarget{method-Element-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ``Element`` instance for ``LRUCache``.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$new(value, creation_time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(ValType): Value that is stored in cache.}

\item{\code{creation_time}}{(datetime.datetime): Time at which cache item was created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Element-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Element-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
