% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{Enum}
\alias{Enum}
\title{Create Enum "like" environments}
\usage{
Enum(..., .class = NULL)
}
\arguments{
\item{...}{(obj): parameters to be create into an Enum like environment}

\item{.class}{(str):}
}
\description{
Create Enum "like" environments
}
\seealso{
Other r_utils: 
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
