% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{evaluate_filter_expression}
\alias{evaluate_filter_expression}
\title{Evaluates model filter with cached model spec value, returns boolean.}
\usage{
evaluate_filter_expression(model_filter, cached_model_value)
}
\arguments{
\item{model_filter}{(ModelFilter): The model filter for evaluation.}

\item{cached_model_value}{(Any): The value in the model manifest/spec that should be used to
evaluate the filter.}
}
\description{
Evaluates model filter with cached model spec value, returns boolean.
}
