% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_notebook_utils.R
\name{extract_framework_task_model}
\alias{extract_framework_task_model}
\title{Parse the model ID, return a tuple framework, task, rest-of-id.}
\usage{
extract_framework_task_model(model_id)
}
\arguments{
\item{model_id}{(str): The model ID for which to extract the framework/task/model.}
}
\description{
Parse the model ID, return a tuple framework, task, rest-of-id.
}
