% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_session.R
\name{file_input}
\alias{file_input}
\title{Amazon SageMaker channel configuration for FILE data sources, used in local mode.}
\description{
Amazon SageMaker channel configuration for FILE data sources, used in local mode.

Amazon SageMaker channel configuration for FILE data sources, used in local mode.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{file_input$new()}}
\item \href{#method-clone}{\code{file_input$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a definition for input data used by an SageMaker training job in local mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{file_input$new(fileUri, content_type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileUri}}{(str):}

\item{\code{content_type}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{file_input$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
