% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\docType{data}
\name{FilterOperators}
\alias{FilterOperators}
\title{Enum class for filter operators for JumpStart models.}
\format{
An object of class \code{FilterOperators} (inherits from \code{Enum}, \code{environment}) of length 4.
}
\usage{
FilterOperators
}
\description{
Enum class for filter operators for JumpStart models.
}
\keyword{datasets}
