% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{format_class}
\alias{format_class}
\title{Format of R6 classes}
\usage{
format_class(self)
}
\arguments{
\item{self}{([R6::R6Class])}
}
\description{
Format of R6 classes
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
