% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{framework_name_from_image}
\alias{framework_name_from_image}
\title{Extract the framework and Python version from the image name.}
\usage{
framework_name_from_image(image_uri)
}
\arguments{
\item{image_uri}{(str): Image URI, which should be one of the following forms:
\itemize{
   \item{legacy: '<account>.dkr.ecr.<region>.amazonaws.com/sagemaker-<fw>-<py_ver>-<device>:<container_version>'}
   \item{legacy: '<account>.dkr.ecr.<region>.amazonaws.com/sagemaker-<fw>-<py_ver>-<device>:<fw_version>-<device>-<py_ver>'}
   \item{current: '<account>.dkr.ecr.<region>.amazonaws.com/sagemaker-<fw>:<fw_version>-<device>-<py_ver>'}
   \item{current: '<account>.dkr.ecr.<region>.amazonaws.com/sagemaker-rl-<fw>:<rl_toolkit><rl_version>-<device>-<py_ver>'}
   \item{current: '<account>.dkr.ecr.<region>.amazonaws.com/<fw>-<image_scope>:<fw_version>-<device>-<py_ver>'}
}}
}
\value{
tuple: A tuple containing:
\itemize{
   \item{str: The framework name}
   \item{str: The Python version}
   \item{str: The image tag}
   \item{str: If the TensorFlow image is script mode}
}
}
\description{
Extract the framework and Python version from the image name.
}
