% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{framework_version_from_tag}
\alias{framework_version_from_tag}
\title{Extract the framework version from the image tag.}
\usage{
framework_version_from_tag(image_tag)
}
\arguments{
\item{image_tag}{(str): Image tag, which should take the form
'<framework_version>-<device>-<py_version>'}
}
\value{
str: The framework version.
}
\description{
Extract the framework version from the image tag.
}
