% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{get_batch_strategy_instance}
\alias{get_batch_strategy_instance}
\title{Return an Instance of :class:`sagemaker.local.data.BatchStrategy` according to `strategy`}
\usage{
get_batch_strategy_instance(strategy, splitter)
}
\arguments{
\item{strategy}{(str): Either 'SingleRecord' or 'MultiRecord'}

\item{splitter}{(:class:`sagemaker.local.data.Splitter): splitter to get the data from.}
}
\value{
:class:`sagemaker.local.data.BatchStrategy`: an Instance of a BatchStrategy
}
\description{
Return an Instance of :class:`sagemaker.local.data.BatchStrategy` according to `strategy`
}
