% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config_value}
\alias{get_config_value}
\title{Placeholder}
\usage{
get_config_value(key_path, config = NULL)
}
\arguments{
\item{key_path}{(str):}

\item{config}{(str):}
}
\description{
Placeholder
}
\seealso{
Other sagemaker_utils: 
\code{\link{.aws_partition}()},
\code{\link{.download_files_under_prefix}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{create_tar_file}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
\keyword{internal}
