% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{get_data_source_instance}
\alias{get_data_source_instance}
\title{Return an Instance of :class:`sagemaker.local.data.DataSource`.}
\usage{
get_data_source_instance(data_source, sagemaker_session)
}
\arguments{
\item{data_source}{(str): a valid URI that points to a data source.}

\item{sagemaker_session}{(:class:`sagemaker.session.Session`): a SageMaker Session to
interact with S3 if required.}
}
\value{
sagemaker.local.data.DataSource: an Instance of a Data Source
}
\description{
The instance can handle the provided data_source URI.
             data_source can be either file:// or s3://
}
