% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{get_formatted_manifest}
\alias{get_formatted_manifest}
\title{Returns formatted manifest dictionary from raw manifest.}
\usage{
get_formatted_manifest(manifest)
}
\arguments{
\item{manifest}{: Placeholder}
}
\description{
Keys are JumpStartVersionedModelId objects, values are
             ``JumpStartModelHeader`` objects
}
