% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{get_jumpstart_base_name_if_jumpstart_model}
\alias{get_jumpstart_base_name_if_jumpstart_model}
\title{Return default JumpStart base name if a URI belongs to JumpStart.}
\usage{
get_jumpstart_base_name_if_jumpstart_model(uris)
}
\arguments{
\item{uris}{(Optional[str]): URI to test for association with JumpStart.}
}
\description{
If no URIs belong to JumpStart, return None.
}
