% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{get_jumpstart_content_bucket}
\alias{get_jumpstart_content_bucket}
\title{Returns regionalized content bucket name for JumpStart.}
\usage{
get_jumpstart_content_bucket(region)
}
\arguments{
\item{region}{(str): AWS region}
}
\description{
Returns regionalized content bucket name for JumpStart.
}
