% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{get_jumpstart_launched_regions_message}
\alias{get_jumpstart_launched_regions_message}
\title{Returns formatted string indicating where JumpStart is launched.}
\usage{
get_jumpstart_launched_regions_message()
}
\description{
Returns formatted string indicating where JumpStart is launched.
}
