% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_model_package_args}
\alias{get_model_package_args}
\title{Get arguments for create_model_package method.}
\usage{
get_model_package_args(
  content_types,
  response_types,
  inference_instances,
  transform_instances,
  model_package_name = NULL,
  model_package_group_name = NULL,
  model_data = NULL,
  image_uri = NULL,
  model_metrics = NULL,
  metadata_properties = NULL,
  marketplace_cert = FALSE,
  approval_status = NULL,
  description = NULL,
  tags = NULL,
  container_def_list = NULL,
  drift_check_baselines = NULL
)
}
\arguments{
\item{content_types}{(list): The supported MIME types for the input data.}

\item{response_types}{(list): The supported MIME types for the output data.}

\item{inference_instances}{(list): A list of the instance types that are used to
generate inferences in real-time.}

\item{transform_instances}{(list): A list of the instance types on which a transformation
job can be run or on which an endpoint can be deployed.}

\item{model_package_name}{(str): Model Package name, exclusive to `model_package_group_name`,
using `model_package_name` makes the Model Package un-versioned (default: None).}

\item{model_package_group_name}{(str): Model Package Group name, exclusive to
`model_package_name`, using `model_package_group_name` makes the Model Package
versioned (default: None).}

\item{model_data}{: Placeholder}

\item{image_uri}{(str): Inference image uri for the container. Model class' self.image will
be used if it is None (default: None).}

\item{model_metrics}{(ModelMetrics): ModelMetrics object (default: None).}

\item{metadata_properties}{(MetadataProperties): MetadataProperties object (default: None).}

\item{marketplace_cert}{(bool): A boolean value indicating if the Model Package is certified
for AWS Marketplace (default: False).}

\item{approval_status}{(str): Model Approval Status, values can be "Approved", "Rejected",
or "PendingManualApproval" (default: "PendingManualApproval").}

\item{description}{(str): Model Package description (default: None).}

\item{tags}{: Placeholder}

\item{container_def_list}{(list): A list of container defintiions.}

\item{drift_check_baselines}{(DriftCheckBaselines): DriftCheckBaselines object (default: None).}
}
\value{
list: A dictionary of method argument names and values.
}
\description{
Get arguments for create_model_package method.
}
