% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{get_mp_parameters}
\alias{get_mp_parameters}
\title{Get the model parallelism parameters provided by the user.}
\usage{
get_mp_parameters(distribution)
}
\arguments{
\item{distribution}{: distribution dictionary defined by the user.}
}
\value{
params: dictionary containing model parallelism parameters
             used for training.
}
\description{
Get the model parallelism parameters provided by the user.
}
