% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{get_sagemaker_version}
\alias{get_sagemaker_version}
\title{Returns sagemaker library version.}
\usage{
get_sagemaker_version()
}
\description{
If the sagemaker library version has not been set, this function
             calls ``parse_sagemaker_version`` to retrieve the version and set
             the constant.
}
