% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_short_version}
\alias{get_short_version}
\title{Return short version in the format of x.x}
\usage{
get_short_version(framework_version)
}
\arguments{
\item{framework_version}{(str): The version string to be shortened.}
}
\value{
str: The short version string
}
\description{
Return short version in the format of x.x
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
