% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{get_splitter_instance}
\alias{get_splitter_instance}
\title{Return an Instance of :class:`sagemaker.local.data.Splitter`.}
\usage{
get_splitter_instance(split_type = NULL)
}
\arguments{
\item{split_type}{(str): either 'Line' or 'RecordIO'. Can be left as None to
signal no data split will happen.}
}
\value{
:class:`sagemaker.local.data.Splitter`: an Instance of a Splitter
}
\description{
The instance returned is according to the specified `split_type`.
}
