% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{HyperparameterValidationMode}
\alias{HyperparameterValidationMode}
\title{Possible modes for validating hyperparameters.}
\format{
An object of class \code{HyperparameterValidationMode} (inherits from \code{Enum}, \code{environment}) of length 3.
}
\usage{
HyperparameterValidationMode
}
\description{
Possible modes for validating hyperparameters.
}
\keyword{datasets}
