% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{is_jumpstart_model_uri}
\alias{is_jumpstart_model_uri}
\title{Returns True if URI corresponds to a JumpStart-hosted model.}
\usage{
is_jumpstart_model_uri(uri)
}
\arguments{
\item{uri}{(Optional[str]): uri for inference/training job.}
}
\description{
Returns True if URI corresponds to a JumpStart-hosted model.
}
