% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{is_list_named}
\alias{is_list_named}
\title{Check if list is named}
\usage{
is_list_named(x)
}
\arguments{
\item{x}{: object}
}
\description{
Check if list is named
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
