% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{is_tarfile}
\alias{is_tarfile}
\title{Check if file is tar archived or not.}
\usage{
is_tarfile(path)
}
\arguments{
\item{path}{A character of filepath to tar archived file.}
}
\description{
Check the magic bytes at offset 257.
             If they match "ustar" including the null terminator, the file is probably a tar.
             \url{https://www.gnu.org/software/tar/manual/html_node/Standard.html}
}
