% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{islistempty}
\alias{islistempty}
\title{Check if list is empty}
\usage{
islistempty(obj)
}
\arguments{
\item{obj}{(list):}
}
\description{
Check if list is empty
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
