% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{IsSubR6Class}
\alias{IsSubR6Class}
\title{Checks is R6 is a sub class}
\usage{
IsSubR6Class(subclass, cls)
}
\arguments{
\item{subclass}{(R6):}

\item{cls}{(R6):}
}
\description{
Checks is R6 is a sub class
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
