% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartDataHolderType}
\alias{JumpStartDataHolderType}
\title{Base class for many JumpStart types.}
\description{
Allows objects to be added to dicts and sets,
             and improves string representation. This class overrides the ``__eq__``
             and ``__hash__`` methods so that different objects with the same attributes/types
             can be compared.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartDataHolderType-format}{\code{JumpStartDataHolderType$format()}}
\item \href{#method-JumpStartDataHolderType-clone}{\code{JumpStartDataHolderType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartDataHolderType-format"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartDataHolderType-format}{}}}
\subsection{Method \code{format()}}{
Returns ``__repr__`` string of object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartDataHolderType$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartDataHolderType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartDataHolderType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartDataHolderType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
