% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartECRSpecs}
\alias{JumpStartECRSpecs}
\title{Data class for JumpStart ECR specs.}
\description{
Data class for JumpStart ECR specs.

Data class for JumpStart ECR specs.
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartECRSpecs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JumpStartECRSpecs$new()}}
\item \href{#method-from_json}{\code{JumpStartECRSpecs$from_json()}}
\item \href{#method-to_json}{\code{JumpStartECRSpecs$to_json()}}
\item \href{#method-clone}{\code{JumpStartECRSpecs$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format">}\href{../../sagemaker.core/html/JumpStartDataHolderType.html#method-format}{\code{sagemaker.core::JumpStartDataHolderType$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a JumpStartECRSpecs object from its json representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartECRSpecs$new(spec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spec}}{(Dict[str, Any]): Dictionary representation of spec.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-from_json}{}}}
\subsection{Method \code{from_json()}}{
Sets fields in object based on json.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartECRSpecs$from_json(json_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_obj}}{(Dict[str, Any]): Dictionary representation of spec.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
Returns json representation of JumpStartECRSpecs object in list format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartECRSpecs$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartECRSpecs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
