% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartEnvironmentVariable}
\alias{JumpStartEnvironmentVariable}
\title{JumpStartEnvironmentVariable class}
\description{
Data class for JumpStart environment variable definitions in the hosting container.
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartEnvironmentVariable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartEnvironmentVariable-new}{\code{JumpStartEnvironmentVariable$new()}}
\item \href{#method-JumpStartEnvironmentVariable-from_json}{\code{JumpStartEnvironmentVariable$from_json()}}
\item \href{#method-JumpStartEnvironmentVariable-to_json}{\code{JumpStartEnvironmentVariable$to_json()}}
\item \href{#method-JumpStartEnvironmentVariable-clone}{\code{JumpStartEnvironmentVariable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format"><a href='../../sagemaker.core/html/JumpStartDataHolderType.html#method-JumpStartDataHolderType-format'><code>sagemaker.core::JumpStartDataHolderType$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartEnvironmentVariable-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartEnvironmentVariable-new}{}}}
\subsection{Method \code{new()}}{
Initializes a JumpStartEnvironmentVariable object from its json representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartEnvironmentVariable$new(spec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spec}}{(Dict[str, Any]): Dictionary representation of environment variable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartEnvironmentVariable-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartEnvironmentVariable-from_json}{}}}
\subsection{Method \code{from_json()}}{
Sets fields in object based on json.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartEnvironmentVariable$from_json(json_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_obj}}{(Dict[str, Any]): Dictionary representation of environment variable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartEnvironmentVariable-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartEnvironmentVariable-to_json}{}}}
\subsection{Method \code{to_json()}}{
Returns json representation of JumpStartEnvironmentVariable object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartEnvironmentVariable$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartEnvironmentVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartEnvironmentVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartEnvironmentVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
