% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_exceptions.R
\name{JumpStartHyperparametersError}
\alias{JumpStartHyperparametersError}
\title{Exception raised for bad hyperparameters of a JumpStart model.}
\description{
Exception raised for bad hyperparameters of a JumpStart model.

Exception raised for bad hyperparameters of a JumpStart model.
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{JumpStartHyperparametersError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartHyperparametersError-clone}{\code{JumpStartHyperparametersError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartHyperparametersError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartHyperparametersError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartHyperparametersError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
