% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartLaunchedRegionInfo}
\alias{JumpStartLaunchedRegionInfo}
\title{Data class for launched region info.}
\description{
Data class for launched region info.

Data class for launched region info.
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartLaunchedRegionInfo}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content_bucket}}{Name of JumpStart s3 content bucket associated with region.}

\item{\code{region_name}}{Name of JumpStart launched region.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartLaunchedRegionInfo-new}{\code{JumpStartLaunchedRegionInfo$new()}}
\item \href{#method-JumpStartLaunchedRegionInfo-clone}{\code{JumpStartLaunchedRegionInfo$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format"><a href='../../sagemaker.core/html/JumpStartDataHolderType.html#method-JumpStartDataHolderType-format'><code>sagemaker.core::JumpStartDataHolderType$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartLaunchedRegionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartLaunchedRegionInfo-new}{}}}
\subsection{Method \code{new()}}{
Instantiates JumpStartLaunchedRegionInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartLaunchedRegionInfo$new(content_bucket, region_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content_bucket}}{(str): Name of JumpStart s3 content bucket associated with region.}

\item{\code{region_name}}{(str): Name of JumpStart launched region.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartLaunchedRegionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartLaunchedRegionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartLaunchedRegionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
