% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartModelHeader}
\alias{JumpStartModelHeader}
\title{Data class JumpStart model header.}
\description{
Data class JumpStart model header.

Data class JumpStart model header.
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartModelHeader}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartModelHeader-new}{\code{JumpStartModelHeader$new()}}
\item \href{#method-JumpStartModelHeader-to_json}{\code{JumpStartModelHeader$to_json()}}
\item \href{#method-JumpStartModelHeader-from_json}{\code{JumpStartModelHeader$from_json()}}
\item \href{#method-JumpStartModelHeader-clone}{\code{JumpStartModelHeader$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format"><a href='../../sagemaker.core/html/JumpStartDataHolderType.html#method-JumpStartDataHolderType-format'><code>sagemaker.core::JumpStartDataHolderType$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelHeader-new}{}}}
\subsection{Method \code{new()}}{
Initializes a JumpStartModelHeader object from its json representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelHeader$new(header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{header}}{(Dict[str, str]): Dictionary representation of header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelHeader-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelHeader-to_json}{}}}
\subsection{Method \code{to_json()}}{
Returns json representation of JumpStartModelHeader object in list format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelHeader$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelHeader-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelHeader-from_json}{}}}
\subsection{Method \code{from_json()}}{
Sets fields in object based on json of header.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelHeader$from_json(json_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_obj}}{(Dict[str, str]): Dictionary representation of header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
