% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_cache.R
\name{JumpStartModelsCache}
\alias{JumpStartModelsCache}
\title{Class that implements a cache for JumpStart models manifests and specs.}
\description{
The manifest and specs associated with JumpStart models provide the information necessary
             for launching JumpStart models from the SageMaker SDK.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartModelsCache-new}{\code{JumpStartModelsCache$new()}}
\item \href{#method-JumpStartModelsCache-set_region}{\code{JumpStartModelsCache$set_region()}}
\item \href{#method-JumpStartModelsCache-get_region}{\code{JumpStartModelsCache$get_region()}}
\item \href{#method-JumpStartModelsCache-set_manifest_file_s3_key}{\code{JumpStartModelsCache$set_manifest_file_s3_key()}}
\item \href{#method-JumpStartModelsCache-get_manifest_file_s3_key}{\code{JumpStartModelsCache$get_manifest_file_s3_key()}}
\item \href{#method-JumpStartModelsCache-set_s3_bucket_name}{\code{JumpStartModelsCache$set_s3_bucket_name()}}
\item \href{#method-JumpStartModelsCache-get_bucket}{\code{JumpStartModelsCache$get_bucket()}}
\item \href{#method-JumpStartModelsCache-get_manifest}{\code{JumpStartModelsCache$get_manifest()}}
\item \href{#method-JumpStartModelsCache-get_header}{\code{JumpStartModelsCache$get_header()}}
\item \href{#method-JumpStartModelsCache-get_specs}{\code{JumpStartModelsCache$get_specs()}}
\item \href{#method-JumpStartModelsCache-clear}{\code{JumpStartModelsCache$clear()}}
\item \href{#method-JumpStartModelsCache-clone}{\code{JumpStartModelsCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``JumpStartModelsCache`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$new(
  region = JUMPSTART_DEFAULT_REGION_NAME(),
  max_s3_cache_items = JUMPSTART_DEFAULT_MAX_S3_CACHE_ITEMS,
  s3_cache_expiration_horizon = JUMPSTART_DEFAULT_S3_CACHE_EXPIRATION_HORIZON,
  max_semantic_version_cache_items = JUMPSTART_DEFAULT_MAX_SEMANTIC_VERSION_CACHE_ITEMS,
 
    semantic_version_cache_expiration_horizon = JUMPSTART_DEFAULT_SEMANTIC_VERSION_CACHE_EXPIRATION_HORIZON,
  manifest_file_s3_key = JUMPSTART_DEFAULT_MANIFEST_FILE_S3_KEY,
  s3_bucket_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{(str): AWS region to associate with cache. Default: region associated
with boto3 session.}

\item{\code{max_s3_cache_items}}{(int): Maximum number of items to store in s3 cache.
Default: 20.}

\item{\code{s3_cache_expiration_horizon}}{(datetime.timedelta): Maximum time to hold
items in s3 cache before invalidation. Default: 6 hours.}

\item{\code{max_semantic_version_cache_items}}{(int): Maximum number of items to store in
semantic version cache. Default: 20.}

\item{\code{semantic_version_cache_expiration_horizon}}{(datetime.timedelta):
Maximum time to hold items in semantic version cache before invalidation.
Default: 6 hours.}

\item{\code{manifest_file_s3_key}}{(str): The key in S3 corresponding to the sdk metadata manifest.}

\item{\code{s3_bucket_name}}{(Optional[str]): S3 bucket to associate with cache.
Default: JumpStart-hosted content bucket for region.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-set_region"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-set_region}{}}}
\subsection{Method \code{set_region()}}{
Set region for cache. Clears cache after new region is set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$set_region(region)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{AWS region to associate with cache.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_region"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_region}{}}}
\subsection{Method \code{get_region()}}{
Return region for cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_region()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-set_manifest_file_s3_key"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-set_manifest_file_s3_key}{}}}
\subsection{Method \code{set_manifest_file_s3_key()}}{
Set manifest file s3 key. Clears cache after new key is set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$set_manifest_file_s3_key(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{(str): The key in S3 corresponding to the sdk metadata manifest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_manifest_file_s3_key"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_manifest_file_s3_key}{}}}
\subsection{Method \code{get_manifest_file_s3_key()}}{
Return manifest file s3 key for cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_manifest_file_s3_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-set_s3_bucket_name"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-set_s3_bucket_name}{}}}
\subsection{Method \code{set_s3_bucket_name()}}{
Set s3 bucket used for cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$set_s3_bucket_name()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_bucket_name}}{(str): S3 bucket to associate with cache.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_bucket"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_bucket}{}}}
\subsection{Method \code{get_bucket()}}{
Return bucket used for cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_bucket()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_manifest"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_manifest}{}}}
\subsection{Method \code{get_manifest()}}{
Return entire JumpStart models manifest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_manifest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_header"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_header}{}}}
\subsection{Method \code{get_header()}}{
Return header for a given JumpStart model ID and semantic version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_header()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{(str): model ID for which to get a header.}

\item{\code{semantic_version_str}}{(str): The semantic version for which to get a
header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-get_specs"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-get_specs}{}}}
\subsection{Method \code{get_specs()}}{
Return specs for a given JumpStart model ID and semantic version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$get_specs()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{(str): model ID for which to get specs.}

\item{\code{semantic_version_str}}{(str): The semantic version for which to get
specs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-clear"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-clear}{}}}
\subsection{Method \code{clear()}}{
Clears the model ID/version and s3 cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelsCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelsCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelsCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
