% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartModelSpecs}
\alias{JumpStartModelSpecs}
\title{JumpStartModelSpecs class}
\description{
Data class JumpStart model specs
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartModelSpecs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartModelSpecs-new}{\code{JumpStartModelSpecs$new()}}
\item \href{#method-JumpStartModelSpecs-from_json}{\code{JumpStartModelSpecs$from_json()}}
\item \href{#method-JumpStartModelSpecs-to_json}{\code{JumpStartModelSpecs$to_json()}}
\item \href{#method-JumpStartModelSpecs-clone}{\code{JumpStartModelSpecs$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format"><a href='../../sagemaker.core/html/JumpStartDataHolderType.html#method-JumpStartDataHolderType-format'><code>sagemaker.core::JumpStartDataHolderType$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelSpecs-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelSpecs-new}{}}}
\subsection{Method \code{new()}}{
Initializes a JumpStartModelSpecs object from its json representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelSpecs$new(spec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spec}}{(Dict[str, Any]): Dictionary representation of spec.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelSpecs-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelSpecs-from_json}{}}}
\subsection{Method \code{from_json()}}{
Sets fields in object based on json of header.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelSpecs$from_json(json_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_obj}}{(Dict[str, Any]): Dictionary representation of spec.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelSpecs-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelSpecs-to_json}{}}}
\subsection{Method \code{to_json()}}{
Returns json representation of JumpStartModelSpecs object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelSpecs$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartModelSpecs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartModelSpecs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartModelSpecs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
