% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\docType{data}
\name{JumpStartS3FileType}
\alias{JumpStartS3FileType}
\title{Type of files published in JumpStart S3 distribution buckets.}
\format{
An object of class \code{JumpStartS3FileType} (inherits from \code{Enum}, \code{environment}) of length 2.
}
\usage{
JumpStartS3FileType
}
\description{
Type of files published in JumpStart S3 distribution buckets.
}
\keyword{datasets}
