% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{JumpStartScriptScope}
\alias{JumpStartScriptScope}
\title{Enum class for JumpStart script scopes.}
\format{
An object of class \code{JumpStartScriptScope} (inherits from \code{Enum}, \code{environment}) of length 2.
}
\usage{
JumpStartScriptScope
}
\description{
Enum class for JumpStart script scopes.
}
\keyword{datasets}
