% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{JumpStartTag}
\alias{JumpStartTag}
\title{Enum class for tag keys to apply to JumpStart models.}
\format{
An object of class \code{JumpStartTag} (inherits from \code{Enum}, \code{environment}) of length 4.
}
\usage{
JumpStartTag
}
\description{
Enum class for tag keys to apply to JumpStart models.
}
\keyword{datasets}
