% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{LineSplitter}
\alias{LineSplitter}
\title{Split records by new line.}
\description{
Split records by new line.

Split records by new line.
}
\section{Super class}{
\code{\link[sagemaker.core:Splitter]{sagemaker.core::Splitter}} -> \code{LineSplitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-split}{\code{LineSplitter$split()}}
\item \href{#method-clone}{\code{LineSplitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
Split a file into records using a specific strategy
             This LineSplitter splits the file on each line break.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineSplitter$split(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): path to the file to split}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list: for the individual records that were split from
             the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineSplitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
