% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_notebook_utils.R
\name{list_jumpstart_tasks}
\alias{list_jumpstart_tasks}
\title{List tasks for JumpStart, and optionally apply filters to result.}
\usage{
list_jumpstart_tasks(
  filter = Constant$new(BooleanValues$`TRUE`),
  region = JUMPSTART_DEFAULT_REGION_NAME()
)
}
\arguments{
\item{filter}{(Union[Operator, str]): Optional. The filter to apply to list tasks. This can be
either an ``Operator`` type filter (e.g. ``And("task == ic", "framework == pytorch")``),
or simply a string filter which will get serialized into an Identity filter.
(e.g. ``"task == ic"``). If this argument is not supplied, all tasks will be listed.
(Default: Constant(BooleanValues$`TRUE`)).}

\item{region}{(str): Optional. The AWS region from which to retrieve JumpStart metadata regarding
models. (Default: JUMPSTART_DEFAULT_REGION_NAME()).}
}
\description{
List tasks for JumpStart, and optionally apply filters to result.
}
