% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{LocalFileDataSource}
\alias{LocalFileDataSource}
\title{LocalFileDataSource class}
\description{
Represents a data source within the local filesystem.
}
\section{Super class}{
\code{\link[sagemaker.core:DataSource]{sagemaker.core::DataSource}} -> \code{LocalFileDataSource}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalFileDataSource$new()}}
\item \href{#method-get_file_list}{\code{LocalFileDataSource$get_file_list()}}
\item \href{#method-get_root_dir}{\code{LocalFileDataSource$get_root_dir()}}
\item \href{#method-clone}{\code{LocalFileDataSource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize LocalFileDataSource class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileDataSource$new(root_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root_path}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_file_list}{}}}
\subsection{Method \code{get_file_list()}}{
Retrieve the list of absolute paths to all the files in this data source.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileDataSource$get_file_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List[str] List of absolute paths.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_root_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_root_dir}{}}}
\subsection{Method \code{get_root_dir()}}{
Retrieve the absolute path to the root directory of this data source.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileDataSource$get_root_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
str: absolute path to the root directory of this data source.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
