% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_session.R
\name{LocalSagemakerClient}
\alias{LocalSagemakerClient}
\title{A SageMakerClient that implements the API calls locally.}
\description{
Used for doing local training and hosting local endpoints. It still needs access to
             a boto client to interact with S3 but it won't perform any SageMaker call.
             Implements the methods with the same signature as the boto SageMakerClient.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalSagemakerClient-new}{\code{LocalSagemakerClient$new()}}
\item \href{#method-LocalSagemakerClient-create_processing_job}{\code{LocalSagemakerClient$create_processing_job()}}
\item \href{#method-LocalSagemakerClient-describe_processing_job}{\code{LocalSagemakerClient$describe_processing_job()}}
\item \href{#method-LocalSagemakerClient-create_training_job}{\code{LocalSagemakerClient$create_training_job()}}
\item \href{#method-LocalSagemakerClient-describe_training_job}{\code{LocalSagemakerClient$describe_training_job()}}
\item \href{#method-LocalSagemakerClient-create_transform_job}{\code{LocalSagemakerClient$create_transform_job()}}
\item \href{#method-LocalSagemakerClient-describe_transform_job}{\code{LocalSagemakerClient$describe_transform_job()}}
\item \href{#method-LocalSagemakerClient-create_model}{\code{LocalSagemakerClient$create_model()}}
\item \href{#method-LocalSagemakerClient-describe_model}{\code{LocalSagemakerClient$describe_model()}}
\item \href{#method-LocalSagemakerClient-describe_endpoint_config}{\code{LocalSagemakerClient$describe_endpoint_config()}}
\item \href{#method-LocalSagemakerClient-create_endpoint_config}{\code{LocalSagemakerClient$create_endpoint_config()}}
\item \href{#method-LocalSagemakerClient-describe_endpoint}{\code{LocalSagemakerClient$describe_endpoint()}}
\item \href{#method-LocalSagemakerClient-create_endpoint}{\code{LocalSagemakerClient$create_endpoint()}}
\item \href{#method-LocalSagemakerClient-update_endpoint}{\code{LocalSagemakerClient$update_endpoint()}}
\item \href{#method-LocalSagemakerClient-delete_endpoint}{\code{LocalSagemakerClient$delete_endpoint()}}
\item \href{#method-LocalSagemakerClient-delete_endpoint_config}{\code{LocalSagemakerClient$delete_endpoint_config()}}
\item \href{#method-LocalSagemakerClient-delete_model}{\code{LocalSagemakerClient$delete_model()}}
\item \href{#method-LocalSagemakerClient-clone}{\code{LocalSagemakerClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize a LocalSageMakerClient.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$new(sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): a session to use to read configurations
from, and use its boto client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_processing_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_processing_job}{}}}
\subsection{Method \code{create_processing_job()}}{
Creates a processing job in Local Mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_processing_job(
  ProcessingJobName,
  AppSpecification,
  ProcessingResources,
  Environment = NULL,
  ProcessingInputs = NULL,
  ProcessingOutputConfig = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProcessingJobName}}{(str): local processing job name.}

\item{\code{AppSpecification}}{(dict): Identifies the container and application to run.}

\item{\code{ProcessingResources}}{(dict): Identifies the resources to use for local processing.}

\item{\code{Environment}}{(dict, optional): Describes the environment variables to pass
to the container. (Default value = None)}

\item{\code{ProcessingInputs}}{(dict, optional): Describes the processing input data.
(Default value = None)}

\item{\code{ProcessingOutputConfig}}{(dict, optional): Describes the processing output
configuration. (Default value = None)}

\item{\code{...}}{: Keyword arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_processing_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_processing_job}{}}}
\subsection{Method \code{describe_processing_job()}}{
Describes a local processing job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_processing_job(ProcessingJobName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProcessingJobName}}{(str): Processing job name to describe.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(dict) DescribeProcessingJob Response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_training_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_training_job}{}}}
\subsection{Method \code{create_training_job()}}{
Create a training job in Local Mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_training_job(
  TrainingJobName,
  AlgorithmSpecification,
  OutputDataConfig,
  ResourceConfig,
  InputDataConfig = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainingJobName}}{(str): local training job name.}

\item{\code{AlgorithmSpecification}}{(dict): Identifies the training algorithm to use.}

\item{\code{OutputDataConfig}}{(dict): Identifies the location where you want to save the results of
model training.}

\item{\code{ResourceConfig}}{(dict): Identifies the resources to use for local model training.}

\item{\code{InputDataConfig}}{(dict, optional): Describes the training dataset and the location where
it is stored. (Default value = None)}

\item{\code{...}}{: Variables added to hyperparameters}

\item{\code{HyperParameters}}{(dict) [optional]: Specifies these algorithm-specific parameters to
influence the quality of the final model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_training_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_training_job}{}}}
\subsection{Method \code{describe_training_job()}}{
Describe a local training job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_training_job(TrainingJobName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainingJobName}}{(str): Training job name to describe.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(dict) DescribeTrainingJob Response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_transform_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_transform_job}{}}}
\subsection{Method \code{create_transform_job()}}{
Create the transform job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_transform_job(
  TransformJobName,
  ModelName,
  TransformInput,
  TransformOutput,
  TransformResources,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TransformJobName}}{(str):}

\item{\code{ModelName}}{(str):}

\item{\code{TransformInput}}{(str):}

\item{\code{TransformOutput}}{(str):}

\item{\code{TransformResources}}{(str):}

\item{\code{...}}{(obj):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_transform_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_transform_job}{}}}
\subsection{Method \code{describe_transform_job()}}{
Describe the transform job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_transform_job(TransformJobName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TransformJobName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_model"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_model}{}}}
\subsection{Method \code{create_model()}}{
Create a Local Model Object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_model(ModelName, PrimaryContainer, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelName}}{(str): the Model Name}

\item{\code{PrimaryContainer}}{(dict): a SageMaker primary container definition}

\item{\code{...}}{(obj):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_model"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_model}{}}}
\subsection{Method \code{describe_model()}}{
Describe the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_model(ModelName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_endpoint_config"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_endpoint_config}{}}}
\subsection{Method \code{describe_endpoint_config()}}{
Describe the endpoint configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_endpoint_config(EndpointConfigName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointConfigName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_endpoint_config"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_endpoint_config}{}}}
\subsection{Method \code{create_endpoint_config()}}{
Create the endpoint configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_endpoint_config(
  EndpointConfigName,
  ProductionVariants,
  Tags = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointConfigName}}{(str):}

\item{\code{ProductionVariants}}{(str):}

\item{\code{Tags}}{: (Default value = NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-describe_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-describe_endpoint}{}}}
\subsection{Method \code{describe_endpoint()}}{
Describe the endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$describe_endpoint(EndpointName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-create_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-create_endpoint}{}}}
\subsection{Method \code{create_endpoint()}}{
Create the endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$create_endpoint(
  EndpointName,
  EndpointConfigName,
  Tags = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointName}}{(str):}

\item{\code{EndpointConfigName}}{(str):}

\item{\code{Tags}}{: (Default value = None)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-update_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-update_endpoint}{}}}
\subsection{Method \code{update_endpoint()}}{
Update the endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$update_endpoint(EndpointName, EndpointConfigName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointName}}{(str):}

\item{\code{EndpointConfigName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-delete_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-delete_endpoint}{}}}
\subsection{Method \code{delete_endpoint()}}{
Delete the endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$delete_endpoint(EndpointName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-delete_endpoint_config"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-delete_endpoint_config}{}}}
\subsection{Method \code{delete_endpoint_config()}}{
Delete the endpoint configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$delete_endpoint_config(EndpointConfigName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EndpointConfigName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-delete_model"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-delete_model}{}}}
\subsection{Method \code{delete_model()}}{
Delete the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$delete_model(ModelName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelName}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
