% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_session.R
\name{LocalSagemakerRuntimeClient}
\alias{LocalSagemakerRuntimeClient}
\title{A SageMaker Runtime client that calls a local endpoint only.}
\description{
A SageMaker Runtime client that calls a local endpoint only.

A SageMaker Runtime client that calls a local endpoint only.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalSagemakerRuntimeClient-new}{\code{LocalSagemakerRuntimeClient$new()}}
\item \href{#method-LocalSagemakerRuntimeClient-invoke_endpoint}{\code{LocalSagemakerRuntimeClient$invoke_endpoint()}}
\item \href{#method-LocalSagemakerRuntimeClient-clone}{\code{LocalSagemakerRuntimeClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerRuntimeClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerRuntimeClient-new}{}}}
\subsection{Method \code{new()}}{
Initializes a LocalSageMakerRuntimeClient.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerRuntimeClient$new(config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{(list): Optional configuration for this client. In particular only
the local port is read.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerRuntimeClient-invoke_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerRuntimeClient-invoke_endpoint}{}}}
\subsection{Method \code{invoke_endpoint()}}{
Invoke the endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerRuntimeClient$invoke_endpoint(
  Body,
  EndpointName,
  ContentType = NULL,
  Accept = NULL,
  CustomAttributes = NULL,
  TargetModel = NULL,
  TargetVariant = NULL,
  InferenceId = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Body}}{: Input data for which you want the model to provide inference.}

\item{\code{EndpointName}}{: The name of the endpoint that you specified when you
created the endpoint using the CreateEndpoint API.}

\item{\code{ContentType}}{: The MIME type of the input data in the request body (Default value = None)}

\item{\code{Accept}}{: The desired MIME type of the inference in the response (Default value = None)}

\item{\code{CustomAttributes}}{: Provides additional information about a request for an inference
submitted to a model hosted at an Amazon SageMaker endpoint (Default value = None)}

\item{\code{TargetModel}}{: The model to request for inference when invoking a multi-model endpoint
(Default value = None)}

\item{\code{TargetVariant}}{: Specify the production variant to send the inference request to when
invoking an endpoint that is running two or more variants (Default value = None)}

\item{\code{InferenceId}}{: If you provide a value, it is added to the captured data when you enable
data capture on the endpoint (Default value = None)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object: Inference for the given input.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSagemakerRuntimeClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSagemakerRuntimeClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSagemakerRuntimeClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
