% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_session.R
\name{LocalSession}
\alias{LocalSession}
\title{A SageMaker ``Session`` class for Local Mode.}
\description{
A SageMaker ``Session`` class for Local Mode.

A SageMaker ``Session`` class for Local Mode.
}
\seealso{
Other Session: 
\code{\link{PawsSession}},
\code{\link{Session}}
}
\concept{Session}
\section{Super class}{
\code{\link[sagemaker.core:Session]{sagemaker.core::Session}} -> \code{LocalSession}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalSession-new}{\code{LocalSession$new()}}
\item \href{#method-LocalSession-logs_for_job}{\code{LocalSession$logs_for_job()}}
\item \href{#method-LocalSession-clone}{\code{LocalSession$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="account_id"><a href='../../sagemaker.core/html/Session.html#method-Session-account_id'><code>sagemaker.core::Session$account_id()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="auto_ml"><a href='../../sagemaker.core/html/Session.html#method-Session-auto_ml'><code>sagemaker.core::Session$auto_ml()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="compile_model"><a href='../../sagemaker.core/html/Session.html#method-Session-compile_model'><code>sagemaker.core::Session$compile_model()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_endpoint"><a href='../../sagemaker.core/html/Session.html#method-Session-create_endpoint'><code>sagemaker.core::Session$create_endpoint()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_endpoint_config"><a href='../../sagemaker.core/html/Session.html#method-Session-create_endpoint_config'><code>sagemaker.core::Session$create_endpoint_config()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_endpoint_config_from_existing"><a href='../../sagemaker.core/html/Session.html#method-Session-create_endpoint_config_from_existing'><code>sagemaker.core::Session$create_endpoint_config_from_existing()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_feature_group"><a href='../../sagemaker.core/html/Session.html#method-Session-create_feature_group'><code>sagemaker.core::Session$create_feature_group()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_model"><a href='../../sagemaker.core/html/Session.html#method-Session-create_model'><code>sagemaker.core::Session$create_model()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_model_from_job"><a href='../../sagemaker.core/html/Session.html#method-Session-create_model_from_job'><code>sagemaker.core::Session$create_model_from_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_model_package_from_algorithm"><a href='../../sagemaker.core/html/Session.html#method-Session-create_model_package_from_algorithm'><code>sagemaker.core::Session$create_model_package_from_algorithm()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_model_package_from_containers"><a href='../../sagemaker.core/html/Session.html#method-Session-create_model_package_from_containers'><code>sagemaker.core::Session$create_model_package_from_containers()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-create_monitoring_schedule'><code>sagemaker.core::Session$create_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="create_tuning_job"><a href='../../sagemaker.core/html/Session.html#method-Session-create_tuning_job'><code>sagemaker.core::Session$create_tuning_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="default_bucket"><a href='../../sagemaker.core/html/Session.html#method-Session-default_bucket'><code>sagemaker.core::Session$default_bucket()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="delete_endpoint"><a href='../../sagemaker.core/html/Session.html#method-Session-delete_endpoint'><code>sagemaker.core::Session$delete_endpoint()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="delete_endpoint_config"><a href='../../sagemaker.core/html/Session.html#method-Session-delete_endpoint_config'><code>sagemaker.core::Session$delete_endpoint_config()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="delete_feature_group"><a href='../../sagemaker.core/html/Session.html#method-Session-delete_feature_group'><code>sagemaker.core::Session$delete_feature_group()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="delete_model"><a href='../../sagemaker.core/html/Session.html#method-Session-delete_model'><code>sagemaker.core::Session$delete_model()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="delete_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-delete_monitoring_schedule'><code>sagemaker.core::Session$delete_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_auto_ml_job"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_auto_ml_job'><code>sagemaker.core::Session$describe_auto_ml_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_feature_group"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_feature_group'><code>sagemaker.core::Session$describe_feature_group()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_model"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_model'><code>sagemaker.core::Session$describe_model()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_monitoring_schedule'><code>sagemaker.core::Session$describe_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_processing_job"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_processing_job'><code>sagemaker.core::Session$describe_processing_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_training_job"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_training_job'><code>sagemaker.core::Session$describe_training_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_transform_job"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_transform_job'><code>sagemaker.core::Session$describe_transform_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="describe_tuning_job"><a href='../../sagemaker.core/html/Session.html#method-Session-describe_tuning_job'><code>sagemaker.core::Session$describe_tuning_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="download_athena_query_result"><a href='../../sagemaker.core/html/Session.html#method-Session-download_athena_query_result'><code>sagemaker.core::Session$download_athena_query_result()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="download_data"><a href='../../sagemaker.core/html/Session.html#method-Session-download_data'><code>sagemaker.core::Session$download_data()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="endpoint_from_job"><a href='../../sagemaker.core/html/Session.html#method-Session-endpoint_from_job'><code>sagemaker.core::Session$endpoint_from_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="endpoint_from_model_data"><a href='../../sagemaker.core/html/Session.html#method-Session-endpoint_from_model_data'><code>sagemaker.core::Session$endpoint_from_model_data()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="endpoint_from_production_variants"><a href='../../sagemaker.core/html/Session.html#method-Session-endpoint_from_production_variants'><code>sagemaker.core::Session$endpoint_from_production_variants()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="expand_role"><a href='../../sagemaker.core/html/Session.html#method-Session-expand_role'><code>sagemaker.core::Session$expand_role()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="format"><a href='../../sagemaker.core/html/Session.html#method-Session-format'><code>sagemaker.core::Session$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="get_caller_identity_arn"><a href='../../sagemaker.core/html/Session.html#method-Session-get_caller_identity_arn'><code>sagemaker.core::Session$get_caller_identity_arn()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="get_query_execution"><a href='../../sagemaker.core/html/Session.html#method-Session-get_query_execution'><code>sagemaker.core::Session$get_query_execution()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="help"><a href='../../sagemaker.core/html/Session.html#method-Session-help'><code>sagemaker.core::Session$help()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="list_candidates"><a href='../../sagemaker.core/html/Session.html#method-Session-list_candidates'><code>sagemaker.core::Session$list_candidates()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="list_monitoring_executions"><a href='../../sagemaker.core/html/Session.html#method-Session-list_monitoring_executions'><code>sagemaker.core::Session$list_monitoring_executions()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="list_monitoring_schedules"><a href='../../sagemaker.core/html/Session.html#method-Session-list_monitoring_schedules'><code>sagemaker.core::Session$list_monitoring_schedules()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="list_s3_files"><a href='../../sagemaker.core/html/Session.html#method-Session-list_s3_files'><code>sagemaker.core::Session$list_s3_files()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="list_tags"><a href='../../sagemaker.core/html/Session.html#method-Session-list_tags'><code>sagemaker.core::Session$list_tags()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="logs_for_auto_ml_job"><a href='../../sagemaker.core/html/Session.html#method-Session-logs_for_auto_ml_job'><code>sagemaker.core::Session$logs_for_auto_ml_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="logs_for_processing_job"><a href='../../sagemaker.core/html/Session.html#method-Session-logs_for_processing_job'><code>sagemaker.core::Session$logs_for_processing_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="logs_for_transform_job"><a href='../../sagemaker.core/html/Session.html#method-Session-logs_for_transform_job'><code>sagemaker.core::Session$logs_for_transform_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="process"><a href='../../sagemaker.core/html/Session.html#method-Session-process'><code>sagemaker.core::Session$process()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="read_s3_file"><a href='../../sagemaker.core/html/Session.html#method-Session-read_s3_file'><code>sagemaker.core::Session$read_s3_file()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="start_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-start_monitoring_schedule'><code>sagemaker.core::Session$start_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="start_query_execution"><a href='../../sagemaker.core/html/Session.html#method-Session-start_query_execution'><code>sagemaker.core::Session$start_query_execution()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="stop_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-stop_monitoring_schedule'><code>sagemaker.core::Session$stop_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="stop_processing_job"><a href='../../sagemaker.core/html/Session.html#method-Session-stop_processing_job'><code>sagemaker.core::Session$stop_processing_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="stop_training_job"><a href='../../sagemaker.core/html/Session.html#method-Session-stop_training_job'><code>sagemaker.core::Session$stop_training_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="stop_transform_job"><a href='../../sagemaker.core/html/Session.html#method-Session-stop_transform_job'><code>sagemaker.core::Session$stop_transform_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="stop_tuning_job"><a href='../../sagemaker.core/html/Session.html#method-Session-stop_tuning_job'><code>sagemaker.core::Session$stop_tuning_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="train"><a href='../../sagemaker.core/html/Session.html#method-Session-train'><code>sagemaker.core::Session$train()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="transform"><a href='../../sagemaker.core/html/Session.html#method-Session-transform'><code>sagemaker.core::Session$transform()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="tune"><a href='../../sagemaker.core/html/Session.html#method-Session-tune'><code>sagemaker.core::Session$tune()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="update_endpoint"><a href='../../sagemaker.core/html/Session.html#method-Session-update_endpoint'><code>sagemaker.core::Session$update_endpoint()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="update_monitoring_schedule"><a href='../../sagemaker.core/html/Session.html#method-Session-update_monitoring_schedule'><code>sagemaker.core::Session$update_monitoring_schedule()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="update_training_job"><a href='../../sagemaker.core/html/Session.html#method-Session-update_training_job'><code>sagemaker.core::Session$update_training_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="upload_data"><a href='../../sagemaker.core/html/Session.html#method-Session-upload_data'><code>sagemaker.core::Session$upload_data()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="upload_string_as_file_body"><a href='../../sagemaker.core/html/Session.html#method-Session-upload_string_as_file_body'><code>sagemaker.core::Session$upload_string_as_file_body()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_athena_query"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_athena_query'><code>sagemaker.core::Session$wait_for_athena_query()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_auto_ml_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_auto_ml_job'><code>sagemaker.core::Session$wait_for_auto_ml_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_compilation_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_compilation_job'><code>sagemaker.core::Session$wait_for_compilation_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_edge_packaging_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_edge_packaging_job'><code>sagemaker.core::Session$wait_for_edge_packaging_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_endpoint"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_endpoint'><code>sagemaker.core::Session$wait_for_endpoint()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_job'><code>sagemaker.core::Session$wait_for_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_model_package"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_model_package'><code>sagemaker.core::Session$wait_for_model_package()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_processing_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_processing_job'><code>sagemaker.core::Session$wait_for_processing_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_transform_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_transform_job'><code>sagemaker.core::Session$wait_for_transform_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="wait_for_tuning_job"><a href='../../sagemaker.core/html/Session.html#method-Session-wait_for_tuning_job'><code>sagemaker.core::Session$wait_for_tuning_job()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Session" data-id="was_processing_job_successful"><a href='../../sagemaker.core/html/Session.html#method-Session-was_processing_job_successful'><code>sagemaker.core::Session$was_processing_job_successful()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSession-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSession-new}{}}}
\subsection{Method \code{new()}}{
This class provides alternative Local Mode implementations for the functionality of
             :class:`~sagemaker.session.Session`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$new(
  paws_session = NULL,
  s3_endpoint_url = NULL,
  disable_local_code = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paws_session}}{(PawsSession): The underlying AWS credentails passed to paws SDK.}

\item{\code{s3_endpoint_url}}{(str): Override the default endpoint URL for Amazon S3, if set
(default: None).}

\item{\code{disable_local_code}}{(bool): Set ``True`` to override the default AWS configuration
chain to disable the ``local.local_code`` setting, which may not be supported for
some SDK features (default: False).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSession-logs_for_job"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSession-logs_for_job}{}}}
\subsection{Method \code{logs_for_job()}}{
A no-op method meant to override the sagemaker client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$logs_for_job(job_name, wait = FALSE, poll = 5, log_type = "All")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_name}}{(str):}

\item{\code{wait}}{(boolean): (Default value = False)}

\item{\code{poll}}{(int): (Default value = 5)}

\item{\code{log_type}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalSession-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalSession-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
