% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{log_stream}
\alias{log_stream}
\title{Format logging streams from aws cloud watch}
\usage{
log_stream(cloudwatchlogs, log_group, stream_name, start_time = 0, skip = 1)
}
\arguments{
\item{cloudwatchlogs}{(str):}

\item{log_group}{(str):}

\item{stream_name}{(str):}

\item{start_time}{(int):}

\item{skip}{(str):}
}
\description{
Format logging streams from aws cloud watch
}
\keyword{internal}
