% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{model_code_key_prefix}
\alias{model_code_key_prefix}
\title{Returns the s3 key prefix for uploading code during model deployment}
\usage{
model_code_key_prefix(code_location_key_prefix, model_name, image)
}
\arguments{
\item{code_location_key_prefix}{(str): the s3 key prefix from code_location}

\item{model_name}{(str): the name of the model}

\item{image}{(str): the image from which a default name can be extracted}
}
\value{
str: the key prefix to be used in uploading code
}
\description{
The location returned is a potential concatenation of 2 parts
             1. code_location_key_prefix if it exists
             2. model_name or a name derived from the image
}
