% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{ModelFramework}
\alias{ModelFramework}
\title{Enum class for JumpStart model framework.}
\format{
An object of class \code{ModelFramework} (inherits from \code{Enum}, \code{environment}) of length 8.
}
\usage{
ModelFramework
}
\description{
The ML framework as referenced in the prefix of the model ID.
             This value does not necessarily correspond to the container name.
}
\keyword{datasets}
