% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{MultiRecordStrategy}
\alias{MultiRecordStrategy}
\title{Feed multiple records at a time for batch inference.}
\description{
Will group up as many records as possible within the payload specified.
}
\section{Super class}{
\code{\link[sagemaker.core:BatchStrategy]{sagemaker.core::BatchStrategy}} -> \code{MultiRecordStrategy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiRecordStrategy-pad}{\code{MultiRecordStrategy$pad()}}
\item \href{#method-MultiRecordStrategy-clone}{\code{MultiRecordStrategy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="BatchStrategy" data-id="initialize"><a href='../../sagemaker.core/html/BatchStrategy.html#method-BatchStrategy-initialize'><code>sagemaker.core::BatchStrategy$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiRecordStrategy-pad"></a>}}
\if{latex}{\out{\hypertarget{method-MultiRecordStrategy-pad}{}}}
\subsection{Method \code{pad()}}{
Group together as many records as possible to fit in the specified size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiRecordStrategy$pad(file, size = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): file path to read the records from.}

\item{\code{size}}{(int): maximum size in MB that each group of records will be
fitted to. passing 0 means unlimited size.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generator of records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiRecordStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiRecordStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiRecordStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
