% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_from_base}
\alias{name_from_base}
\title{Append a timestamp to the provided string.}
\usage{
name_from_base(base, max_length = 63, short = FALSE)
}
\arguments{
\item{base}{(str): String used as prefix to generate the unique name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}

\item{short}{(bool): Whether or not to use a truncated timestamp (default: False).}
}
\value{
str: Input parameter with appended timestamp.
}
\description{
This function assures that the total length of the resulting string is
             not longer than the specified max length, trimming the input parameter if
             necessary.
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
