% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_from_image}
\alias{name_from_image}
\title{Create a training job name based on the image name and a timestamp.}
\usage{
name_from_image(image, max_length = 63L)
}
\arguments{
\item{image}{(str): Image name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}
}
\value{
str: Training job name using the algorithm from the image name and a
       timestamp.
}
\description{
Create a training job name based on the image name and a timestamp.
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
