% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{NoneSplitter}
\alias{NoneSplitter}
\title{NoneSplitter class}
\description{
Does not split records, essentially reads the whole file.
}
\section{Super class}{
\code{\link[sagemaker.core:Splitter]{sagemaker.core::Splitter}} -> \code{NoneSplitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-split}{\code{NoneSplitter$split()}}
\item \href{#method-clone}{\code{NoneSplitter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
Split a file into records using a specific strategy.
             For this NoneSplitter there is no actual split happening and the file
             is returned as a whole.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoneSplitter$split(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): path to the file to split}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generator for the individual records that were split from
             the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NoneSplitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
