% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Not}
\alias{Not}
\title{Not operator class for filtering JumpStart content.}
\description{
Not operator class for filtering JumpStart content.

Not operator class for filtering JumpStart content.
}
\section{Super classes}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{\link[sagemaker.core:Operator]{sagemaker.core::Operator}} -> \code{Not}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Not-new}{\code{Not$new()}}
\item \href{#method-Not-eval}{\code{Not$eval()}}
\item \href{#method-Not-clone}{\code{Not$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format"><a href='../../sagemaker.core/html/Operand.html#method-Operand-format'><code>sagemaker.core::Operand$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand"><a href='../../sagemaker.core/html/Operand.html#method-Operand-validate_operand'><code>sagemaker.core::Operand$validate_operand()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Not-new"></a>}}
\if{latex}{\out{\hypertarget{method-Not-new}{}}}
\subsection{Method \code{new()}}{
Instantiates Not object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Not$new(operand)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operand}}{(Operand): Operand for Not-ing.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Not-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Not-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Not$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Not-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Not-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Not$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
