% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Operand}
\alias{Operand}
\title{Operand class for filtering JumpStart content.}
\description{
Operand class for filtering JumpStart content.

Operand class for filtering JumpStart content.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{resolved_value}}{Getter method for resolved_value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Operand-new}{\code{Operand$new()}}
\item \href{#method-Operand-validate_operand}{\code{Operand$validate_operand()}}
\item \href{#method-Operand-eval}{\code{Operand$eval()}}
\item \href{#method-Operand-format}{\code{Operand$format()}}
\item \href{#method-Operand-clone}{\code{Operand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operand-new"></a>}}
\if{latex}{\out{\hypertarget{method-Operand-new}{}}}
\subsection{Method \code{new()}}{
Initialize Operand Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operand$new(unresolved_value, resolved_value = BooleanValues$UNEVALUATED)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unresolved_value}}{(Any): The unresolved value of the operator.}

\item{\code{resolved_value}}{(BooleanValues): The resolved value of the operator.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operand-validate_operand"></a>}}
\if{latex}{\out{\hypertarget{method-Operand-validate_operand}{}}}
\subsection{Method \code{validate_operand()}}{
Validate operand and return ``Operand`` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operand$validate_operand(operand)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operand}}{(Any): The operand to validate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operand-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Operand-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operand.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operand$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operand-format"></a>}}
\if{latex}{\out{\hypertarget{method-Operand-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operand$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Operand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
