% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Operator}
\alias{Operator}
\title{Operator class for filtering JumpStart content.}
\description{
An operator in this case corresponds to an operand that is also an operation.
             For example, given the expression ``(True or True) and True``,
             ``(True or True)`` is an operand to an ``And`` expression, but is also itself an
             operator. ``(True or True) and True`` would also be considered an operator.
}
\section{Super class}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{Operator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Operator-new}{\code{Operator$new()}}
\item \href{#method-Operator-eval}{\code{Operator$eval()}}
\item \href{#method-Operator-clone}{\code{Operator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format"><a href='../../sagemaker.core/html/Operand.html#method-Operand-format'><code>sagemaker.core::Operand$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand"><a href='../../sagemaker.core/html/Operand.html#method-Operand-validate_operand'><code>sagemaker.core::Operand$validate_operand()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-new}{}}}
\subsection{Method \code{new()}}{
Initializes ``Operator`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$new(
  resolved_value = BooleanValues$UNEVALUATED,
  unresolved_value = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resolved_value}}{(BooleanValues): Optional. The resolved value of the operator.
(Default: BooleanValues.UNEVALUATED).}

\item{\code{unresolved_value}}{(Any): Optional. The unresolved value of the operator.
(Default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
