% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Or}
\alias{Or}
\title{Or operator class for filtering JumpStart content.}
\description{
Or operator class for filtering JumpStart content.

Or operator class for filtering JumpStart content.
}
\section{Super classes}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{\link[sagemaker.core:Operator]{sagemaker.core::Operator}} -> \code{Or}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Or-new}{\code{Or$new()}}
\item \href{#method-Or-eval}{\code{Or$eval()}}
\item \href{#method-Or-clone}{\code{Or$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format"><a href='../../sagemaker.core/html/Operand.html#method-Operand-format'><code>sagemaker.core::Operand$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand"><a href='../../sagemaker.core/html/Operand.html#method-Operand-validate_operand'><code>sagemaker.core::Operand$validate_operand()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Or-new"></a>}}
\if{latex}{\out{\hypertarget{method-Or-new}{}}}
\subsection{Method \code{new()}}{
Instantiates Or object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Or$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(Operand): Operand for Or-ing.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Or-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Or-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Or$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Or-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Or-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Or$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
