% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{parse_filter_string}
\alias{parse_filter_string}
\title{Parse filter string and return a serialized ``ModelFilter`` object.}
\usage{
parse_filter_string(filter_string)
}
\arguments{
\item{filter_string}{(str): The filter string to be serialized to an object.}
}
\description{
Parse filter string and return a serialized ``ModelFilter`` object.
}
