% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{parse_s3}
\alias{parse_s3}
\alias{split_s3_uri}
\alias{parse_s3_url}
\title{split s3 uri}
\usage{
split_s3_uri(uri)

parse_s3_url(uri)
}
\arguments{
\item{uri}{(str): s3 uri to split into bucket and key}
}
\description{
split s3 uri
}
