% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{parse_sagemaker_version}
\alias{parse_sagemaker_version}
\title{Returns sagemaker library version. This should only be called once.}
\usage{
parse_sagemaker_version()
}
\description{
Function reads ``__version__`` variable in ``sagemaker`` module.
             In order to maintain compatibility with the ``packaging.version``
             library, versions with fewer than 2, or more than 3, periods are rejected.
             All versions that cannot be parsed with ``packaging.version`` are also
             rejected.
}
