% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws_session.R
\name{PawsSession}
\alias{PawsSession}
\title{PawsSession Class create connection to AWS utilizing paws.}
\description{
A session stores configuration state and allows you to create paws service
             clients.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aws_access_key_id}}{aws access key}

\item{\code{aws_secret_access_key}}{aws secret access key}

\item{\code{aws_session_token}}{aws session token}

\item{\code{region_name}}{Default region when creating new connections}

\item{\code{profile_name}}{The name of a profile to use.}

\item{\code{endpoint}}{The complete URL to use for the constructed client.}

\item{\code{credentials}}{Formatted aws credentials to pass to paws objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PawsSession$new()}}
\item \href{#method-client}{\code{PawsSession$client()}}
\item \href{#method-format}{\code{PawsSession$format()}}
\item \href{#method-clone}{\code{PawsSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize PawsSession class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsSession$new(
  aws_access_key_id = NULL,
  aws_secret_access_key = NULL,
  aws_session_token = NULL,
  region_name = NULL,
  profile_name = NULL,
  endpoint = NULL,
  config = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_access_key_id}}{(str): AWS access key ID}

\item{\code{aws_secret_access_key}}{(str): AWS secret access key}

\item{\code{aws_session_token}}{(str): AWS temporary session token}

\item{\code{region_name}}{(str): Default region when creating new connections}

\item{\code{profile_name}}{(str): The name of a profile to use. If not given, then the default profile is used.}

\item{\code{endpoint}}{(str): The complete URL to use for the constructed client.}

\item{\code{config}}{(list): Optional paws configuration of credentials, endpoint, and/or region.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-client"></a>}}
\if{latex}{\out{\hypertarget{method-client}{}}}
\subsection{Method \code{client()}}{
Create a low-level service client by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsSession$client(service_name, config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service_name}}{(str): The name of a service, e.g. 's3' or 'ec2'.
A list of available services can be found \url{https://paws-r.github.io/docs/}}

\item{\code{config}}{(list): Optional paws configuration of credentials, endpoint, and/or region.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsSession$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
