% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{pkg_method}
\alias{pkg_method}
\title{Get methods from other packages}
\usage{
pkg_method(fun, pkg)
}
\arguments{
\item{fun}{function to export}

\item{pkg}{package to method from}
}
\description{
This function allows to use soft dependencies.
}
\keyword{internal}
