% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{python_deprecation_warning}
\alias{python_deprecation_warning}
\title{Raise warning for deprecated python versions}
\usage{
python_deprecation_warning(framework, latest_supported_version)
}
\arguments{
\item{framework}{(str): model framework}

\item{latest_supported_version}{(str): latest supported version}
}
\description{
Raise warning for deprecated python versions
}
