% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiutils_base_types.R
\name{Record}
\alias{Record}
\title{Record class}
\description{
A boto based Active Record class based on convention of CRUD operations.
}
\section{Super class}{
\code{\link[sagemaker.core:ApiObject]{sagemaker.core::ApiObject}} -> \code{Record}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Record-new}{\code{Record$new()}}
\item \href{#method-Record-with_paws}{\code{Record$with_paws()}}
\item \href{#method-Record-clone}{\code{Record$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="format"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-format'><code>sagemaker.core::ApiObject$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="from_paws"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-from_paws'><code>sagemaker.core::ApiObject$from_paws()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="to_paws"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-to_paws'><code>sagemaker.core::ApiObject$to_paws()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Record-new"></a>}}
\if{latex}{\out{\hypertarget{method-Record-new}{}}}
\subsection{Method \code{new()}}{
Init Record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$new(sagemaker_session = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}

\item{\code{...}}{parameters passed to `R6` class `ApiObject`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Record-with_paws"></a>}}
\if{latex}{\out{\hypertarget{method-Record-with_paws}{}}}
\subsection{Method \code{with_paws()}}{
Update this ApiObject with a paws response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$with_paws(paws_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paws_list}}{(dict): A dictionary of a paws response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Record-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Record-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
