% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{RecordIOSplitter}
\alias{RecordIOSplitter}
\title{Split using Amazon Recordio.}
\description{
Not useful for string content.
}
\section{Super class}{
\code{\link[sagemaker.core:Splitter]{sagemaker.core::Splitter}} -> \code{RecordIOSplitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordIOSplitter-split}{\code{RecordIOSplitter$split()}}
\item \href{#method-RecordIOSplitter-clone}{\code{RecordIOSplitter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordIOSplitter-split"></a>}}
\if{latex}{\out{\hypertarget{method-RecordIOSplitter-split}{}}}
\subsection{Method \code{split()}}{
Split a file into records using a specific strategy
             This RecordIOSplitter splits the data into individual RecordIO
             records.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordIOSplitter$split(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): path to the file to split}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generator for the individual records that were split from
             the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordIOSplitter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecordIOSplitter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordIOSplitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
