% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_utils.R
\name{recursive_copy}
\alias{recursive_copy}
\title{A wrapper around distutils.dir_util.copy_tree.}
\usage{
recursive_copy(source, destination)
}
\arguments{
\item{source}{(str): source path}

\item{destination}{(str): destination path}
}
\description{
This won't throw any exception when the source directory does not exist.
}
