% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{remove_arg}
\alias{remove_arg}
\title{Checks if the deprecated argument is populated.}
\usage{
remove_arg(name, arg = NULL)
}
\arguments{
\item{name}{(str): name of deprecated argument}

\item{arg}{(str): the argument to check}
}
\description{
Raises warning, if not None.
}
