% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{removed_kwargs}
\alias{removed_kwargs}
\title{Checks if the deprecated argument is in kwargs}
\usage{
removed_kwargs(name, kwargs)
}
\arguments{
\item{name}{(str): name of deprecated argument}

\item{kwargs}{(str): keyword arguments dict}
}
\description{
Raises warning, if present.
}
