% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{removed_warning}
\alias{removed_warning}
\title{Raise a warning for a no-op in sagemaker>=2}
\usage{
removed_warning(phrase, sdk_version = NULL)
}
\arguments{
\item{phrase}{(str): the prefix phrase of the warning message.}

\item{sdk_version}{(str): the sdk version of removal of support.}
}
\description{
Raise a warning for a no-op in sagemaker>=2
}
