% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{renamed_warning}
\alias{renamed_warning}
\title{Raise a warning for a rename in sagemaker>=2}
\usage{
renamed_warning(phrase)
}
\arguments{
\item{phrase}{(str): the prefix phrase of the warning message.}
}
\description{
Raise a warning for a rename in sagemaker>=2
}
