% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repack_model}
\alias{repack_model}
\title{Unpack model tarball and creates a new model tarball with the provided
             code script.}
\usage{
repack_model(
  inference_script,
  source_directory,
  dependencies,
  model_uri,
  repacked_model_uri,
  sagemaker_session,
  kms_key = NULL
)
}
\arguments{
\item{inference_script}{(str): path or basename of the inference script that
will be packed into the model}

\item{source_directory}{(str): path including all the files that will be packed
into the model}

\item{dependencies}{(list[str]): A list of paths to directories (absolute or
relative) with any additional libraries that will be exported to the}

\item{model_uri}{(str): S3 or file system location of the original model tar}

\item{repacked_model_uri}{(str): path or file system location where the new
model will be saved}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}

\item{kms_key}{(str): KMS key ARN for encrypting the repacked model file}
}
\value{
str: path to the new packed model
}
\description{
This function does the following: - uncompresses model tarball from S3 or
             local system into a temp folder - replaces the inference code from the model
             with the new code provided - compresses the new model tarball and saves it
             in S3 or local file system
}
