% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{retry_api_call}
\alias{retry_api_call}
\title{If api call fails retry call}
\usage{
retry_api_call(expr, retry = 5)
}
\arguments{
\item{expr}{(code): AWS code to rety}

\item{retry}{(int): number of retries}
}
\description{
If api call fails retry call
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{rsplit}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
