% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{rsplit}
\alias{rsplit}
\title{Split string from the right}
\usage{
rsplit(str, separator = "\\\\.", maxsplit)
}
\arguments{
\item{str}{: string to be split}

\item{separator}{(str): Method splits string starting from the right (default `\\.`)}

\item{maxsplit}{(number): The maxsplit defines the maximum number of splits.}
}
\description{
Split string from the right
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{split_str}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
