% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{sagemaker_log_layout}
\alias{sagemaker_log_layout}
\title{Set format for SageMaker}
\usage{
sagemaker_log_layout(
  log_fmt = "\%L [\%t] \%m",
  timestamp_fmt = "\%Y-\%m-\%d \%H:\%M:\%OS",
  log_cols = log_colours()
)
}
\arguments{
\item{log_fmt}{(str): a character scalar containing format tokens}

\item{timestamp_fmt}{(str): a character scalar.}

\item{log_cols}{(list): a named list of functions (like the ones provided by the package crayon)}
}
\description{
Set format for SageMaker
}
\keyword{internal}
