% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{SagemakerError}
\alias{SagemakerError}
\alias{NotImplementedError}
\alias{ValueError}
\alias{TypeError}
\alias{UnexpectedStatusError}
\alias{RuntimeError}
\alias{AttributeError}
\alias{DataNotFoundError}
\alias{UnknownServiceError}
\alias{KeyError}
\title{SagemakerError}
\description{
Create Sagemaker error format.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SagemakerError-new}{\code{SagemakerError$new()}}
\item \href{#method-SagemakerError-clone}{\code{SagemakerError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SagemakerError-new"></a>}}
\if{latex}{\out{\hypertarget{method-SagemakerError-new}{}}}
\subsection{Method \code{new()}}{
Initialize SagemakerError class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SagemakerError$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(character): message to be outputted in error.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SagemakerError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SagemakerError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SagemakerError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{NotImplementedError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotImplementedError-clone}{\code{NotImplementedError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotImplementedError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotImplementedError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotImplementedError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{ValueError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValueError-clone}{\code{ValueError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValueError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{TypeError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TypeError-clone}{\code{TypeError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TypeError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TypeError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{\link[sagemaker.core:ValueError]{sagemaker.core::ValueError}} -> \code{UnexpectedStatusError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnexpectedStatusError-new}{\code{UnexpectedStatusError$new()}}
\item \href{#method-UnexpectedStatusError-clone}{\code{UnexpectedStatusError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnexpectedStatusError-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnexpectedStatusError-new}{}}}
\subsection{Method \code{new()}}{
Initialize UnexpectedStatusError class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnexpectedStatusError$new(..., allowed_statuses, actual_status)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(character): message to be outputted in error.}

\item{\code{allowed_statuses}}{(character): allowed status from sagemaker}

\item{\code{actual_status}}{(character): returning status from aws sagemaker}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnexpectedStatusError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnexpectedStatusError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnexpectedStatusError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{RuntimeError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RuntimeError-clone}{\code{RuntimeError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RuntimeError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RuntimeError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RuntimeError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{AttributeError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttributeError-clone}{\code{AttributeError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{DataNotFoundError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataNotFoundError-clone}{\code{DataNotFoundError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataNotFoundError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataNotFoundError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataNotFoundError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{\link[sagemaker.core:DataNotFoundError]{sagemaker.core::DataNotFoundError}} -> \code{UnknownServiceError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnknownServiceError-clone}{\code{UnknownServiceError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnknownServiceError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnknownServiceError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnknownServiceError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[sagemaker.core:SagemakerError]{sagemaker.core::SagemakerError}} -> \code{KeyError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KeyError-clone}{\code{KeyError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="SagemakerError" data-id="initialize"><a href='../../sagemaker.core/html/SagemakerError.html#method-SagemakerError-initialize'><code>sagemaker.core::SagemakerError$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeyError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KeyError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeyError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
