% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secondary_training_status_changed}
\alias{secondary_training_status_changed}
\title{Returns true if training job's secondary status message has changed.}
\usage{
secondary_training_status_changed(
  current_job_description = NULL,
  prev_job_description = NULL
)
}
\arguments{
\item{current_job_description}{(str): Current job description, returned from DescribeTrainingJob call.}

\item{prev_job_description}{(str): Previous job description, returned from DescribeTrainingJob call.}
}
\value{
boolean: Whether the secondary status message of a training job changed
             or not.
}
\description{
Returns true if training job's secondary status message has changed.
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
