% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secondary_training_status_message}
\alias{secondary_training_status_message}
\title{Returns a string contains last modified time and the secondary training
             job status message.}
\usage{
secondary_training_status_message(
  job_description = NULL,
  prev_description = NULL
)
}
\arguments{
\item{job_description}{(str): Returned response from DescribeTrainingJob call}

\item{prev_description}{(str): Previous job description from DescribeTrainingJob call}
}
\value{
str: Job status string to be printed.
}
\description{
Returns a string contains last modified time and the secondary training
             job status message.
}
\seealso{
Other sagemaker_utils: 
\code{\link{.aws_partition}()},
\code{\link{.download_files_under_prefix}()},
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{create_tar_file}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_config_value}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
