% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_settings.R
\name{SessionSettings}
\alias{SessionSettings}
\title{Optional container class for settings to apply to a SageMaker session.}
\description{
Optional container class for settings to apply to a SageMaker session.

Optional container class for settings to apply to a SageMaker session.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{encrypt_repacked_artifacts}}{Return True if repacked artifacts at rest in S3 should be encrypted by default.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SessionSettings$new()}}
\item \href{#method-format}{\code{SessionSettings$format()}}
\item \href{#method-clone}{\code{SessionSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ``SessionSettings`` of a SageMaker ``Session``.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionSettings$new(encrypt_repacked_artifacts = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{encrypt_repacked_artifacts}}{(bool): Flag to indicate whether to encrypt the artifacts
at rest in S3 using the default AWS managed KMS key for S3 when a custom KMS key
is not provided (Default: \code{TRUE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionSettings$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
