% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\docType{data}
\name{SpecialSupportedFilterKeys}
\alias{SpecialSupportedFilterKeys}
\title{Enum class for special supported filter keys.}
\format{
An object of class \code{SpecialSupportedFilterKeys} (inherits from \code{Enum}, \code{environment}) of length 3.
}
\usage{
SpecialSupportedFilterKeys
}
\description{
Enum class for special supported filter keys.
}
\keyword{datasets}
