% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{split_str}
\alias{split_str}
\title{split string}
\usage{
split_str(str, split = ",")
}
\arguments{
\item{str}{(str): string to split}

\item{split}{(str): string used for splitting.}
}
\description{
split string
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{write_bin}()}
}
\concept{r_utils}
\keyword{internal}
