% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{tag_key_in_array}
\alias{tag_key_in_array}
\title{Returns True if ``tag_key`` is in the ``tag_array``.}
\usage{
tag_key_in_array(tag_key, tag_array)
}
\arguments{
\item{tag_key}{(str): the tag key to check if it's already in the ``tag_array``.}

\item{tag_array}{(List[Dict[str, str]]): array of tags to check for ``tag_key``.}
}
\description{
Returns True if ``tag_key`` is in the ``tag_array``.
}
