% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{tar_and_upload_dir}
\alias{tar_and_upload_dir}
\title{Package source files and uploads a compress tar file to S3.}
\usage{
tar_and_upload_dir(
  sagemaker_session,
  bucket,
  s3_key_prefix,
  script,
  directory = NULL,
  dependencies = NULL,
  kms_key = NULL
)
}
\arguments{
\item{sagemaker_session}{(sagemaker.Session): sagemaker_session session used to access S3.}

\item{bucket}{(str): S3 bucket to which the compressed file is uploaded.}

\item{s3_key_prefix}{(str): Prefix for the S3 key.}

\item{script}{(str): Script filename or path.}

\item{directory}{(str): Optional. Directory containing the source file. If it
starts with "s3://", no action is taken.}

\item{dependencies}{(List[str]): Optional. A list of paths to directories
(absolute or relative) containing additional libraries that will be
copied into /opt/ml/lib}

\item{kms_key}{(str): Optional. KMS key ID used to upload objects to the bucket
(default: None).}
}
\value{
sagemaker.fw_utils.UserCode: An object with the S3 bucket and key (S3 prefix) and
              script name.
}
\description{
Package source files and upload a compress tar file to S3. The S3
             location will be ``s3://<bucket>/s3_key_prefix/sourcedir.tar.gz``.
             If directory is an S3 URI, an UploadedCode object will be returned, but
             nothing will be uploaded to S3 (this allow reuse of code already in S3).
             If directory is None, the script will be added to the archive at
             ``./<basename of script>``.
             If directory is not None, the (recursive) contents of the directory will
             be added to the archive. directory is treated as the base path of the
             archive, and the script name is assumed to be a filename or relative path
             inside the directory.
}
