% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tar_subdir}
\alias{tar_subdir}
\title{tar archive file and directories}
\usage{
tar_subdir(tarfile, src, compress = "gzip", ...)
}
\arguments{
\item{tarfile}{The pathname of the tar file.}

\item{src}{file or sub directory to be tar archived.}

\item{compress}{character string giving the type of compression to be used (default gzip).}

\item{...}{extra parameters to be passes to \code{\link[utils]{tar}}}
}
\description{
tar archive file and directories
}
