% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{TransformInput}
\alias{TransformInput}
\title{TransformInput}
\description{
Create a class containing all the parameters.
             It can be used when calling ``sagemaker.transformer.Transformer.transform()``
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Place holder}

\item{\code{data_type}}{Place holder}

\item{\code{content_type}}{Place holder}

\item{\code{compression_type}}{Place holder}

\item{\code{split_type}}{Place holder}

\item{\code{input_filter}}{Place holder}

\item{\code{output_filter}}{Place holder}

\item{\code{join_source}}{Place holder}

\item{\code{model_client_config}}{Place holder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TransformInput$new()}}
\item \href{#method-format}{\code{TransformInput$format()}}
\item \href{#method-clone}{\code{TransformInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize TransformInput class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformInput$new(
  data = NULL,
  data_type = "S3Prefix",
  content_type = NULL,
  compression_type = NULL,
  split_type = NULL,
  input_filter = NULL,
  output_filter = NULL,
  join_source = NULL,
  model_client_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(str): Place holder}

\item{\code{data_type}}{(str): Place holder}

\item{\code{content_type}}{(str): Place holder}

\item{\code{compression_type}}{(str): Place holder}

\item{\code{split_type}}{(str): Place holder}

\item{\code{input_filter}}{(str): Place holder}

\item{\code{output_filter}}{(str): Place holder}

\item{\code{join_source}}{(str): Place holder}

\item{\code{model_client_config}}{(str): Place holder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformInput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransformInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
