% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_name_from_base}
\alias{unique_name_from_base}
\title{Create a unique name from base str}
\usage{
unique_name_from_base(base, max_length = 63)
}
\arguments{
\item{base}{(str): String used as prefix to generate the unique name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}
}
\value{
str: Input parameter with appended timestamp.
}
\description{
Create a unique name from base str
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_from_name}()},
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{common_variables}},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()}
}
\concept{sagemaker_utils}
