% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_utils.R
\name{update_inference_tags_with_jumpstart_training_tags}
\alias{update_inference_tags_with_jumpstart_training_tags}
\title{Updates the tags for the ``sagemaker.model.Model.deploy`` command with any JumpStart tags.}
\usage{
update_inference_tags_with_jumpstart_training_tags(
  inference_tags,
  training_tags
)
}
\arguments{
\item{inference_tags}{(Optional[List[Dict[str, str]]]): Custom tags to appy to inference job.}

\item{training_tags}{(Optional[List[Dict[str, str]]]): Tags from training job.}
}
\description{
Updates the tags for the ``sagemaker.model.Model.deploy`` command with any JumpStart tags.
}
