% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\docType{data}
\name{UploadedCode}
\alias{UploadedCode}
\title{sagemaker.fw_utils.UserCode: An object containing the S3 prefix and script name.}
\format{
An object of class \code{list} of length 2.
}
\usage{
UploadedCode
}
\description{
This is for the source code used for the entry point with an ``Estimator``. It can be
             instantiated with positional or keyword arguments.
}
\keyword{internal}
