% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{validate_mp_config}
\alias{validate_mp_config}
\title{Validate the configuration dictionary for model parallelism.}
\usage{
validate_mp_config(config)
}
\arguments{
\item{config}{(list): Dictionary holding configuration keys and values.}
}
\description{
Validate the configuration dictionary for model parallelism.
}
