% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{validate_smdistributed}
\alias{validate_smdistributed}
\title{Check if smdistributed strategy is correctly invoked by the user.}
\usage{
validate_smdistributed(
  instance_type,
  framework_name,
  framework_version,
  py_version,
  distribution,
  image_uri = NULL
)
}
\arguments{
\item{instance_type}{(str): A string representing the type of training instance selected.}

\item{framework_name}{(str): A string representing the name of framework selected.}

\item{framework_version}{(str): A string representing the framework version selected.}

\item{py_version}{(str): A string representing the python version selected.}

\item{distribution}{(dict): A dictionary with information to enable distributed training.
(Defaults to None if distributed training is not enabled.)}

\item{image_uri}{(str): A string representing a Docker image URI.}
}
\description{
Currently, two strategies are supported: `dataparallel` or `modelparallel`.
             Validate if the user requested strategy is supported.
             Currently, only one strategy can be specified at a time. Validate if the user has requested
             more than one strategy simultaneously.
             Validate if the smdistributed dict arg is syntactically correct.
             Additionally, perform strategy-specific validations.
}
