% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{validate_source_dir}
\alias{validate_source_dir}
\title{Validate that the source directory exists and it contains the user script}
\usage{
validate_source_dir(script, directory)
}
\arguments{
\item{script}{(str): Script filename.}

\item{directory}{(str): Directory containing the source file.}
}
\description{
Validate that the source directory exists and it contains the user script
}
