% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{validate_version_or_image_args}
\alias{validate_version_or_image_args}
\title{Checks if version or image arguments are specified.}
\usage{
validate_version_or_image_args(framework_version, py_version, image_uri)
}
\arguments{
\item{framework_version}{(str): The version of the framework.}

\item{py_version}{(str): The version of Python.}

\item{image_uri}{(str): The URI of the image.}
}
\description{
Validates framework and model arguments to enforce version or image specification.
}
