% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{VariableScope}
\alias{VariableScope}
\title{Possible value of the ``scope`` attribute for a hyperparameter or environment variable.}
\format{
An object of class \code{VariableScope} (inherits from \code{Enum}, \code{environment}) of length 2.
}
\usage{
VariableScope
}
\description{
Used for hosting environment variables and training hyperparameters.
}
\keyword{datasets}
