% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_enums.R
\docType{data}
\name{VariableTypes}
\alias{VariableTypes}
\title{Possible types for hyperparameters and environment variables.}
\format{
An object of class \code{VariableTypes} (inherits from \code{Enum}, \code{environment}) of length 4.
}
\usage{
VariableTypes
}
\description{
Possible types for hyperparameters and environment variables.
}
\keyword{datasets}
