% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\docType{data}
\name{vpc_configuration_env}
\alias{vpc_configuration_env}
\title{Vpc Configuration environment list}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 4.
}
\usage{
vpc_configuration_env
}
\description{
Contains default vpc configurations.
}
\seealso{
Other vpc_utils: 
\code{\link{vpc_from_list}()},
\code{\link{vpc_sanitize}()},
\code{\link{vpc_to_list}()}
}
\concept{vpc_utils}
\keyword{internal}
