% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\name{vpc_from_list}
\alias{vpc_from_list}
\title{Extracts subnets and security group ids as lists from a VpcConfig dict}
\usage{
vpc_from_list(vpc_config, do_sanitize = FALSE)
}
\arguments{
\item{vpc_config}{(list): a VpcConfig list containing 'Subnets' and
SecurityGroupIds'}

\item{do_sanitize}{(bool): whether to sanitize the VpcConfig dict before
extracting values}
}
\value{
list as (subnets, security_group_ids) If vpc_config parameter
             is None, returns (None, None)
}
\description{
Extracts subnets and security group ids as lists from a VpcConfig dict
}
\seealso{
Other vpc_utils: 
\code{\link{vpc_sanitize}()},
\code{\link{vpc_to_list}()}
}
\concept{vpc_utils}
