% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\name{vpc_sanitize}
\alias{vpc_sanitize}
\title{Checks VpcConfig}
\usage{
vpc_sanitize(vpc_config = NULL)
}
\arguments{
\item{vpc_config}{(list): a VpcConfig dict containing 'Subnets' and
SecurityGroupIds'}
}
\value{
A valid VpcConfig dict containing only 'Subnets' and 'SecurityGroupIds'
             from the vpc_config parameter If vpc_config parameter is None, returns
             None
}
\description{
Checks that an instance of VpcConfig has the expected keys and values,
             removes unexpected keys, and raises ValueErrors if any expectations are
             violated
}
