% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\name{vpc_to_list}
\alias{vpc_to_list}
\title{Convert subnet and security groups in to vpc list}
\usage{
vpc_to_list(subnets, security_group_ids)
}
\arguments{
\item{subnets}{(list): list of subnet IDs to use in VpcConfig}

\item{security_group_ids}{(list): list of security group IDs to use in
VpcConfig}
}
\value{
A VpcConfig dict containing keys 'Subnets' and 'SecurityGroupIds' If
             either or both parameters are None, returns None
}
\description{
Prepares a VpcConfig list containing keys 'Subnets' and
             SecurityGroupIds' This is the dict format expected by SageMaker
             CreateTrainingJob and CreateModel APIs See
             \url{https://docs.aws.amazon.com/sagemaker/latest/dg/API_VpcConfig.html}
}
