% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_utils.R
\name{warn_if_parameter_server_with_multi_gpu}
\alias{warn_if_parameter_server_with_multi_gpu}
\title{Warn the user that training will not fully leverage all the GPU cores}
\usage{
warn_if_parameter_server_with_multi_gpu(training_instance_type, distribution)
}
\arguments{
\item{training_instance_type}{(str): A string representing the type of training instance selected.}

\item{distribution}{(dict): A dictionary with information to enable distributed training.
(Defaults to None if distributed training is not enabled.).}
}
\description{
Warn the user that training will not fully leverage all the GPU
             cores if parameter server is enabled and a multi-GPU instance is selected.
             Distributed training with the default parameter server setup doesn't
             support multi-GPU instances.
}
