% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{write_bin}
\alias{write_bin}
\title{Write large raw connections in chunks}
\usage{
write_bin(obj, filename, chunk_size = 2^31 - 2)
}
\arguments{
\item{obj}{(raw): raw connection or raw vector}

\item{filename}{(str):}

\item{chunk_size}{(int):}
}
\description{
Write large raw connections in chunks
}
\seealso{
Other r_utils: 
\code{\link{Enum}()},
\code{\link{IsSubR6Class}()},
\code{\link{cls_help}()},
\code{\link{format_class}()},
\code{\link{is_list_named}()},
\code{\link{is_tarfile}()},
\code{\link{islistempty}()},
\code{\link{pkg_method}()},
\code{\link{retry_api_call}()},
\code{\link{rsplit}()},
\code{\link{split_str}()}
}
\concept{r_utils}
\keyword{internal}
