% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_common.R
\name{write_matrix_to_dense_tensor}
\alias{write_matrix_to_dense_tensor}
\title{write_matrix_to_dense_tensor}
\usage{
write_matrix_to_dense_tensor(file, array, labels = NULL)
}
\arguments{
\item{file}{(str): file location}

\item{array}{(array):}

\item{labels}{(str):}
}
\description{
Write matrix to dense tensor file.
}
