% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_common.R
\name{write_spmatrix_to_sparse_tensor}
\alias{write_spmatrix_to_sparse_tensor}
\title{write_spmatrix_to_sparse_tensor}
\usage{
write_spmatrix_to_sparse_tensor(file, array, labels = NULL)
}
\arguments{
\item{file}{(str): file location}

\item{array}{(array):}

\item{labels}{(str):}
}
\description{
Writes a Matrix sparse matrix to a sparse tensor
}
