% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{And}
\alias{And}
\title{And operator class for filtering JumpStart content.}
\description{
And operator class for filtering JumpStart content.

And operator class for filtering JumpStart content.
}
\section{Super classes}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{\link[sagemaker.core:Operator]{sagemaker.core::Operator}} -> \code{And}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-And-new}{\code{And$new()}}
\item \href{#method-And-eval}{\code{And$eval()}}
\item \href{#method-And-clone}{\code{And$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format"><a href='../../sagemaker.core/html/Operand.html#method-Operand-format'><code>sagemaker.core::Operand$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand"><a href='../../sagemaker.core/html/Operand.html#method-Operand-validate_operand'><code>sagemaker.core::Operand$validate_operand()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-And-new"></a>}}
\if{latex}{\out{\hypertarget{method-And-new}{}}}
\subsection{Method \code{new()}}{
Instantiates And object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{And$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(Operand): Operand for And-ing.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-And-eval"></a>}}
\if{latex}{\out{\hypertarget{method-And-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{And$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-And-clone"></a>}}
\if{latex}{\out{\hypertarget{method-And-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{And$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
