% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_from_name}
\alias{base_from_name}
\title{Extract the base name of the resource name (for use with future resource name generation).}
\usage{
base_from_name(name)
}
\arguments{
\item{name}{(str): The resource name.}
}
\value{
str: The base name, as extracted from the resource name.
}
\description{
This function looks for timestamps that match the ones produced by
             :func:`~sagemaker.utils.name_from_base`.
}
\seealso{
Other sagemaker_utils: 
\code{\link{base_name_from_image}()},
\code{\link{build_dict}()},
\code{\link{download_file}()},
\code{\link{download_folder}()},
\code{\link{get_short_version}()},
\code{\link{name_from_base}()},
\code{\link{name_from_image}()},
\code{\link{regional_hostname}()},
\code{\link{repack_model}()},
\code{\link{retries}()},
\code{\link{sagemaker_short_timestamp}()},
\code{\link{sagemaker_timestamp}()},
\code{\link{secondary_training_status_changed}()},
\code{\link{secondary_training_status_message}()},
\code{\link{sts_regional_endpoint}()},
\code{\link{unique_name_from_base}()}
}
\concept{sagemaker_utils}
