% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_from_name}
\alias{base_from_name}
\title{Extract the base name of the resource name (for use with future resource name generation).}
\usage{
base_from_name(name)
}
\arguments{
\item{name}{(str): The resource name.}
}
\value{
str: The base name, as extracted from the resource name.
}
\description{
This function looks for timestamps that match the ones produced by
             :func:`~sagemaker.utils.name_from_base`.
}
