% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_name_from_image}
\alias{base_name_from_image}
\title{Extract the base name of the image to use as the 'algorithm name' for the job.}
\usage{
base_name_from_image(image)
}
\arguments{
\item{image}{(str): Image name.}
}
\value{
str: Algorithm name, as extracted from the image name.
}
\description{
Extract the base name of the image to use as the 'algorithm name' for the job.
}
