% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_dict}
\alias{build_dict}
\title{Return a dict of key and value pair if value is not None, otherwise return an empty dict.}
\usage{
build_dict(key, value = NULL)
}
\arguments{
\item{key}{(str): input key}

\item{value}{(str): input value}
}
\value{
dict: dict of key and value or an empty dict.
}
\description{
Return a dict of key and value pair if value is not None, otherwise return an empty dict.
}
