% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tar_file}
\alias{create_tar_file}
\title{Create a tar file containing all the source_files}
\usage{
create_tar_file(source_files, target = NULL)
}
\arguments{
\item{source_files}{(str): vector of file paths that will be contained in the tar file}

\item{target}{(str):}
}
\value{
(str): path to created tar file
}
\description{
Create a tar file containing all the source_files
}
\keyword{internal}
