% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.aws_partition}
\alias{.aws_partition}
\title{Given a region name (ex: "cn-north-1"), return the corresponding aws partition ("aws-cn").}
\usage{
.aws_partition(region)
}
\value{
str: partition corresponding to the region name passed in.
}
\description{
region (str): The region name for which to return the corresponding partition.
}
\keyword{internal}
