% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.download_files_under_prefix}
\alias{.download_files_under_prefix}
\title{Download all S3 files which match the given prefix}
\usage{
.download_files_under_prefix(bucket_name, prefix, target, s3)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{prefix}{(str): S3 prefix within the bucket that will be downloaded}

\item{target}{(str): destination path where the downloaded items will be placed}

\item{s3}{(paws::s3): S3 resource}
}
\description{
Download all S3 files which match the given prefix
}
\keyword{internal}
