% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a Single File from S3 into a local path}
\usage{
download_file(bucket_name, path, target, sagemaker_session)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{path}{(str): file path within the bucket}

\item{target}{(str): destination directory for the downloaded file.}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}
}
\description{
Download a Single File from S3 into a local path
}
