% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_folder}
\alias{download_folder}
\title{Download a folder from S3 to a local path}
\usage{
download_folder(bucket_name, prefix, target, sagemaker_session)
}
\arguments{
\item{bucket_name}{(str): S3 bucket name}

\item{prefix}{(str): S3 prefix within the bucket that will be downloaded. Can
be a single file.}

\item{target}{(str): destination path where the downloaded items will be placed}

\item{sagemaker_session}{(sagemaker.session.Session): a sagemaker session to
interact with S3.}
}
\description{
Download a folder from S3 to a local path
}
