% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{Enum}
\alias{Enum}
\title{Create Enum "like" environments}
\usage{
Enum(..., .class = NULL)
}
\arguments{
\item{...}{(obj): parameters to be create into an Enum like environment}

\item{.class}{(str):}
}
\description{
Create Enum "like" environments
}
\keyword{internal}
