% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config_value}
\alias{get_config_value}
\title{Placeholder}
\usage{
get_config_value(key_path, config = NULL)
}
\arguments{
\item{key_path}{(str):}

\item{config}{(str):}
}
\description{
Placeholder
}
\keyword{internal}
