% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_short_version}
\alias{get_short_version}
\title{Return short version in the format of x.x}
\usage{
get_short_version(framework_version)
}
\arguments{
\item{framework_version}{(str): The version string to be shortened.}
}
\value{
str: The short version string
}
\description{
Return short version in the format of x.x
}
