% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_filters.R
\name{Identity}
\alias{Identity}
\title{Identity operator class for filtering JumpStart content.}
\description{
Identity operator class for filtering JumpStart content.

Identity operator class for filtering JumpStart content.
}
\section{Super classes}{
\code{\link[sagemaker.core:Operand]{sagemaker.core::Operand}} -> \code{\link[sagemaker.core:Operator]{sagemaker.core::Operator}} -> \code{Identity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Identity$new()}}
\item \href{#method-eval}{\code{Identity$eval()}}
\item \href{#method-clone}{\code{Identity$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="format">}\href{../../sagemaker.core/html/Operand.html#method-format}{\code{sagemaker.core::Operand$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="Operand" data-id="validate_operand">}\href{../../sagemaker.core/html/Operand.html#method-validate_operand}{\code{sagemaker.core::Operand$validate_operand()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiates Identity object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Identity$new(operand)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operand}}{(Union[Operand, str]): Operand for identity operation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates operator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Identity$eval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Identity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
