% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{IsSubR6Class}
\alias{IsSubR6Class}
\title{Checks is R6 is a sub class}
\usage{
IsSubR6Class(subclass, cls)
}
\arguments{
\item{subclass}{(R6):}

\item{cls}{(R6):}
}
\description{
Checks is R6 is a sub class
}
\keyword{internal}
