% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart_types.R
\name{JumpStartVersionedModelId}
\alias{JumpStartVersionedModelId}
\title{JumpStartVersionedModelId class}
\description{
Data class for versioned model IDs.
}
\section{Super class}{
\code{\link[sagemaker.core:JumpStartDataHolderType]{sagemaker.core::JumpStartDataHolderType}} -> \code{JumpStartVersionedModelId}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JumpStartVersionedModelId-new}{\code{JumpStartVersionedModelId$new()}}
\item \href{#method-JumpStartVersionedModelId-clone}{\code{JumpStartVersionedModelId$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="JumpStartDataHolderType" data-id="format"><a href='../../sagemaker.core/html/JumpStartDataHolderType.html#method-JumpStartDataHolderType-format'><code>sagemaker.core::JumpStartDataHolderType$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartVersionedModelId-new"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartVersionedModelId-new}{}}}
\subsection{Method \code{new()}}{
Instantiates JumpStartVersionedModelId object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartVersionedModelId$new(model_id, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{(str): JumpStart model ID.}

\item{\code{version}}{(str): JumpStart model version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JumpStartVersionedModelId-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JumpStartVersionedModelId-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JumpStartVersionedModelId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
