% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_session.R
\name{LocalSession}
\alias{LocalSession}
\title{A SageMaker ``Session`` class for Local Mode.}
\description{
A SageMaker ``Session`` class for Local Mode.

A SageMaker ``Session`` class for Local Mode.
}
\section{Super class}{
\code{\link[R6sagemaker.common:Session]{R6sagemaker.common::Session}} -> \code{LocalSession}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalSession$new()}}
\item \href{#method-logs_for_job}{\code{LocalSession$logs_for_job()}}
\item \href{#method-clone}{\code{LocalSession$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="auto_ml">}\href{../../R6sagemaker.common/html/Session.html#method-auto_ml}{\code{R6sagemaker.common::Session$auto_ml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="compile_model">}\href{../../R6sagemaker.common/html/Session.html#method-compile_model}{\code{R6sagemaker.common::Session$compile_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_endpoint">}\href{../../R6sagemaker.common/html/Session.html#method-create_endpoint}{\code{R6sagemaker.common::Session$create_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_endpoint_config">}\href{../../R6sagemaker.common/html/Session.html#method-create_endpoint_config}{\code{R6sagemaker.common::Session$create_endpoint_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_endpoint_config_from_existing">}\href{../../R6sagemaker.common/html/Session.html#method-create_endpoint_config_from_existing}{\code{R6sagemaker.common::Session$create_endpoint_config_from_existing()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_model">}\href{../../R6sagemaker.common/html/Session.html#method-create_model}{\code{R6sagemaker.common::Session$create_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_model_from_job">}\href{../../R6sagemaker.common/html/Session.html#method-create_model_from_job}{\code{R6sagemaker.common::Session$create_model_from_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_model_package_from_algorithm">}\href{../../R6sagemaker.common/html/Session.html#method-create_model_package_from_algorithm}{\code{R6sagemaker.common::Session$create_model_package_from_algorithm()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_model_package_from_containers">}\href{../../R6sagemaker.common/html/Session.html#method-create_model_package_from_containers}{\code{R6sagemaker.common::Session$create_model_package_from_containers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-create_monitoring_schedule}{\code{R6sagemaker.common::Session$create_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="create_tuning_job">}\href{../../R6sagemaker.common/html/Session.html#method-create_tuning_job}{\code{R6sagemaker.common::Session$create_tuning_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="default_bucket">}\href{../../R6sagemaker.common/html/Session.html#method-default_bucket}{\code{R6sagemaker.common::Session$default_bucket()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="delete_endpoint">}\href{../../R6sagemaker.common/html/Session.html#method-delete_endpoint}{\code{R6sagemaker.common::Session$delete_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="delete_endpoint_config">}\href{../../R6sagemaker.common/html/Session.html#method-delete_endpoint_config}{\code{R6sagemaker.common::Session$delete_endpoint_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="delete_model">}\href{../../R6sagemaker.common/html/Session.html#method-delete_model}{\code{R6sagemaker.common::Session$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="delete_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-delete_monitoring_schedule}{\code{R6sagemaker.common::Session$delete_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_auto_ml_job">}\href{../../R6sagemaker.common/html/Session.html#method-describe_auto_ml_job}{\code{R6sagemaker.common::Session$describe_auto_ml_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-describe_monitoring_schedule}{\code{R6sagemaker.common::Session$describe_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_processing_job">}\href{../../R6sagemaker.common/html/Session.html#method-describe_processing_job}{\code{R6sagemaker.common::Session$describe_processing_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_training_job">}\href{../../R6sagemaker.common/html/Session.html#method-describe_training_job}{\code{R6sagemaker.common::Session$describe_training_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_transform_job">}\href{../../R6sagemaker.common/html/Session.html#method-describe_transform_job}{\code{R6sagemaker.common::Session$describe_transform_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="describe_tuning_job">}\href{../../R6sagemaker.common/html/Session.html#method-describe_tuning_job}{\code{R6sagemaker.common::Session$describe_tuning_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="download_athena_query_result">}\href{../../R6sagemaker.common/html/Session.html#method-download_athena_query_result}{\code{R6sagemaker.common::Session$download_athena_query_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="download_data">}\href{../../R6sagemaker.common/html/Session.html#method-download_data}{\code{R6sagemaker.common::Session$download_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="endpoint_from_job">}\href{../../R6sagemaker.common/html/Session.html#method-endpoint_from_job}{\code{R6sagemaker.common::Session$endpoint_from_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="endpoint_from_model_data">}\href{../../R6sagemaker.common/html/Session.html#method-endpoint_from_model_data}{\code{R6sagemaker.common::Session$endpoint_from_model_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="endpoint_from_production_variants">}\href{../../R6sagemaker.common/html/Session.html#method-endpoint_from_production_variants}{\code{R6sagemaker.common::Session$endpoint_from_production_variants()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="expand_role">}\href{../../R6sagemaker.common/html/Session.html#method-expand_role}{\code{R6sagemaker.common::Session$expand_role()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="format">}\href{../../R6sagemaker.common/html/Session.html#method-format}{\code{R6sagemaker.common::Session$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="get_caller_identity_arn">}\href{../../R6sagemaker.common/html/Session.html#method-get_caller_identity_arn}{\code{R6sagemaker.common::Session$get_caller_identity_arn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="get_query_execution">}\href{../../R6sagemaker.common/html/Session.html#method-get_query_execution}{\code{R6sagemaker.common::Session$get_query_execution()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="list_candidates">}\href{../../R6sagemaker.common/html/Session.html#method-list_candidates}{\code{R6sagemaker.common::Session$list_candidates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="list_monitoring_executions">}\href{../../R6sagemaker.common/html/Session.html#method-list_monitoring_executions}{\code{R6sagemaker.common::Session$list_monitoring_executions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="list_monitoring_schedules">}\href{../../R6sagemaker.common/html/Session.html#method-list_monitoring_schedules}{\code{R6sagemaker.common::Session$list_monitoring_schedules()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="list_s3_files">}\href{../../R6sagemaker.common/html/Session.html#method-list_s3_files}{\code{R6sagemaker.common::Session$list_s3_files()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="list_tags">}\href{../../R6sagemaker.common/html/Session.html#method-list_tags}{\code{R6sagemaker.common::Session$list_tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="logs_for_auto_ml_job">}\href{../../R6sagemaker.common/html/Session.html#method-logs_for_auto_ml_job}{\code{R6sagemaker.common::Session$logs_for_auto_ml_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="logs_for_processing_job">}\href{../../R6sagemaker.common/html/Session.html#method-logs_for_processing_job}{\code{R6sagemaker.common::Session$logs_for_processing_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="logs_for_transform_job">}\href{../../R6sagemaker.common/html/Session.html#method-logs_for_transform_job}{\code{R6sagemaker.common::Session$logs_for_transform_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="process">}\href{../../R6sagemaker.common/html/Session.html#method-process}{\code{R6sagemaker.common::Session$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="read_s3_file">}\href{../../R6sagemaker.common/html/Session.html#method-read_s3_file}{\code{R6sagemaker.common::Session$read_s3_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="start_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-start_monitoring_schedule}{\code{R6sagemaker.common::Session$start_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="start_query_execution">}\href{../../R6sagemaker.common/html/Session.html#method-start_query_execution}{\code{R6sagemaker.common::Session$start_query_execution()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="stop_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-stop_monitoring_schedule}{\code{R6sagemaker.common::Session$stop_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="stop_processing_job">}\href{../../R6sagemaker.common/html/Session.html#method-stop_processing_job}{\code{R6sagemaker.common::Session$stop_processing_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="stop_training_job">}\href{../../R6sagemaker.common/html/Session.html#method-stop_training_job}{\code{R6sagemaker.common::Session$stop_training_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="stop_transform_job">}\href{../../R6sagemaker.common/html/Session.html#method-stop_transform_job}{\code{R6sagemaker.common::Session$stop_transform_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="stop_tuning_job">}\href{../../R6sagemaker.common/html/Session.html#method-stop_tuning_job}{\code{R6sagemaker.common::Session$stop_tuning_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="train">}\href{../../R6sagemaker.common/html/Session.html#method-train}{\code{R6sagemaker.common::Session$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="transform">}\href{../../R6sagemaker.common/html/Session.html#method-transform}{\code{R6sagemaker.common::Session$transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="tune">}\href{../../R6sagemaker.common/html/Session.html#method-tune}{\code{R6sagemaker.common::Session$tune()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="update_endpoint">}\href{../../R6sagemaker.common/html/Session.html#method-update_endpoint}{\code{R6sagemaker.common::Session$update_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="update_monitoring_schedule">}\href{../../R6sagemaker.common/html/Session.html#method-update_monitoring_schedule}{\code{R6sagemaker.common::Session$update_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="upload_data">}\href{../../R6sagemaker.common/html/Session.html#method-upload_data}{\code{R6sagemaker.common::Session$upload_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="upload_string_as_file_body">}\href{../../R6sagemaker.common/html/Session.html#method-upload_string_as_file_body}{\code{R6sagemaker.common::Session$upload_string_as_file_body()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_athena_query">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_athena_query}{\code{R6sagemaker.common::Session$wait_for_athena_query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_auto_ml_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_auto_ml_job}{\code{R6sagemaker.common::Session$wait_for_auto_ml_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_compilation_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_compilation_job}{\code{R6sagemaker.common::Session$wait_for_compilation_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_edge_packaging_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_edge_packaging_job}{\code{R6sagemaker.common::Session$wait_for_edge_packaging_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_endpoint">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_endpoint}{\code{R6sagemaker.common::Session$wait_for_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_job}{\code{R6sagemaker.common::Session$wait_for_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_model_package">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_model_package}{\code{R6sagemaker.common::Session$wait_for_model_package()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_processing_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_processing_job}{\code{R6sagemaker.common::Session$wait_for_processing_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_transform_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_transform_job}{\code{R6sagemaker.common::Session$wait_for_transform_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="wait_for_tuning_job">}\href{../../R6sagemaker.common/html/Session.html#method-wait_for_tuning_job}{\code{R6sagemaker.common::Session$wait_for_tuning_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="Session" data-id="was_processing_job_successful">}\href{../../R6sagemaker.common/html/Session.html#method-was_processing_job_successful}{\code{R6sagemaker.common::Session$was_processing_job_successful()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This class provides alternative Local Mode implementations for the functionality of
             :class:`~sagemaker.session.Session`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$new(
  paws_credentials = NULL,
  s3_endpoint_url = NULL,
  disable_local_code = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paws_credentials}}{(PawsCredentials): The underlying AWS credentails passed to paws SDK.}

\item{\code{s3_endpoint_url}}{(str): Override the default endpoint URL for Amazon S3, if set
(default: None).}

\item{\code{disable_local_code}}{(bool): Set ``True`` to override the default AWS configuration
chain to disable the ``local.local_code`` setting, which may not be supported for
some SDK features (default: False).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logs_for_job"></a>}}
\if{latex}{\out{\hypertarget{method-logs_for_job}{}}}
\subsection{Method \code{logs_for_job()}}{
A no-op method meant to override the sagemaker client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$logs_for_job(job_name, wait = FALSE, poll = 5, log_type = "All")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_name}}{(str):}

\item{\code{wait}}{(boolean): (Default value = False)}

\item{\code{poll}}{(int): (Default value = 5)}

\item{\code{log_type}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
