% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_from_base}
\alias{name_from_base}
\title{Append a timestamp to the provided string.}
\usage{
name_from_base(base, max_length = 63, short = FALSE)
}
\arguments{
\item{base}{(str): String used as prefix to generate the unique name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}

\item{short}{(bool): Whether or not to use a truncated timestamp (default: False).}
}
\value{
str: Input parameter with appended timestamp.
}
\description{
This function assures that the total length of the resulting string is
             not longer than the specified max length, trimming the input parameter if
             necessary.
}
