% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_from_image}
\alias{name_from_image}
\title{Create a training job name based on the image name and a timestamp.}
\usage{
name_from_image(image, max_length = 63L)
}
\arguments{
\item{image}{(str): Image name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}
}
\value{
str: Training job name using the algorithm from the image name and a
       timestamp.
}
\description{
Create a training job name based on the image name and a timestamp.
}
