% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{production_variant}
\alias{production_variant}
\title{Create a production variant description suitable for use in a ``ProductionVariant`` list.}
\usage{
production_variant(
  model_name,
  instance_type,
  initial_instance_count = 1,
  variant_name = "AllTraffic",
  initial_weight = 1,
  accelerator_type = NULL
)
}
\arguments{
\item{model_name}{(str): The name of the SageMaker model this production variant references.}

\item{instance_type}{(str): The EC2 instance type for this production variant. For example,
ml.c4.8xlarge'.}

\item{initial_instance_count}{(int): The initial instance count for this production variant
(default: 1).}

\item{variant_name}{(string): The ``VariantName`` of this production variant
(default: 'AllTraffic').}

\item{initial_weight}{(int): The relative ``InitialVariantWeight`` of this production variant
(default: 1).}

\item{accelerator_type}{(str): Type of Elastic Inference accelerator for this production variant.
For example, 'ml.eia1.medium'.
For more information: \url{https://docs.aws.amazon.com/sagemaker/latest/dg/ei.html}}
}
\value{
dict[str, str]: An SageMaker ``ProductionVariant`` description
}
\description{
This is also part of a ``CreateEndpointConfig`` request.
}
