% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regional_hostname}
\alias{regional_hostname}
\title{Get the AWS endpoint specific for the given region.}
\usage{
regional_hostname(service_name, region)
}
\arguments{
\item{service_name}{(str): Name of the service to resolve an endpoint for (e.g., s3)}

\item{region}{(str): AWS region name}
}
\value{
str: AWS STS regional endpoint
}
\description{
We need this function because the AWS SDK does not yet honor
             the ``region_name`` parameter when creating an AWS STS client.
             For the list of regional endpoints, see
             \url{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html#id_credentials_region-endpoints}.
}
