% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{removed_warning}
\alias{removed_warning}
\title{Raise a warning for a no-op in sagemaker>=2}
\usage{
removed_warning(phrase)
}
\arguments{
\item{phrase}{(str): the prefix phrase of the warning message.}
}
\description{
Raise a warning for a no-op in sagemaker>=2
}
