% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retries}
\alias{retries}
\title{Retries until max retry count is reached.}
\usage{
retries(max_retry_count, exception_message_prefix, seconds_to_sleep = 2)
}
\arguments{
\item{max_retry_count}{(int): The retry count.}

\item{exception_message_prefix}{(str): The message to include in the exception on failure.}

\item{seconds_to_sleep}{(int): The number of seconds to sleep between executions.}
}
\description{
Retries until max retry count is reached.
}
