% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{retry_api_call}
\alias{retry_api_call}
\title{If api call fails retry call}
\usage{
retry_api_call(expr, retry = 5)
}
\arguments{
\item{expr}{(code): AWS code to rety}

\item{retry}{(int): number of retries}
}
\description{
If api call fails retry call
}
\keyword{internal}
