% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{rsplit}
\alias{rsplit}
\title{Split string from the right}
\usage{
rsplit(str, separator = "\\\\.", maxsplit)
}
\arguments{
\item{str}{: string to be split}

\item{separator}{(str): Method splits string starting from the right (default `\\.`)}

\item{maxsplit}{(number): The maxsplit defines the maximum number of splits.}
}
\description{
Split string from the right
}
