% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sagemaker_short_timestamp}
\alias{sagemaker_short_timestamp}
\title{Return a timestamp that is relatively short in length}
\usage{
sagemaker_short_timestamp()
}
\description{
Return a timestamp that is relatively short in length
}
