% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secondary_training_status_changed}
\alias{secondary_training_status_changed}
\title{Returns true if training job's secondary status message has changed.}
\usage{
secondary_training_status_changed(
  current_job_description = NULL,
  prev_job_description = NULL
)
}
\arguments{
\item{current_job_description}{(str): Current job description, returned from DescribeTrainingJob call.}

\item{prev_job_description}{(str): Previous job description, returned from DescribeTrainingJob call.}
}
\value{
boolean: Whether the secondary status message of a training job changed
             or not.
}
\description{
Returns true if training job's secondary status message has changed.
}
