% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secondary_training_status_message}
\alias{secondary_training_status_message}
\title{Returns a string contains last modified time and the secondary training
             job status message.}
\usage{
secondary_training_status_message(
  job_description = NULL,
  prev_description = NULL
)
}
\arguments{
\item{job_description}{(str): Returned response from DescribeTrainingJob call}

\item{prev_description}{(str): Previous job description from DescribeTrainingJob call}
}
\value{
str: Job status string to be printed.
}
\description{
Returns a string contains last modified time and the secondary training
             job status message.
}
