% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{split_str}
\alias{split_str}
\title{split string}
\usage{
split_str(str, split = ",")
}
\arguments{
\item{str}{(str): string to split}

\item{split}{(str): string used for splitting.}
}
\description{
split string
}
\keyword{internal}
