% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{Splitter}
\alias{Splitter}
\title{Splitter class}
\description{
Splitter class

Splitter class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-split}{\code{Splitter$split()}}
\item \href{#method-clone}{\code{Splitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
Split a file into records using a specific strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Splitter$split(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(str): path to the file to split}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
generator for the individual records that were split from the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Splitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
