% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sts_regional_endpoint}
\alias{sts_regional_endpoint}
\title{Get the AWS STS endpoint specific for the given region.}
\usage{
sts_regional_endpoint(region)
}
\arguments{
\item{region}{(str): AWS region name}
}
\value{
str: AWS STS regional endpoint
}
\description{
We need this function because the AWS SDK does not yet honor
             the ``region_name`` parameter when creating an AWS STS client.
             For the list of regional endpoints, see
             \url{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html#id_credentials_region-endpoints}.
}
