% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_name_from_base}
\alias{unique_name_from_base}
\title{Create a unique name from base str}
\usage{
unique_name_from_base(base, max_length = 63)
}
\arguments{
\item{base}{(str): String used as prefix to generate the unique name.}

\item{max_length}{(int): Maximum length for the resulting string (default: 63).}
}
\value{
str: Input parameter with appended timestamp.
}
\description{
Create a unique name from base str
}
