% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\docType{data}
\name{vpc_configuration_env}
\alias{vpc_configuration_env}
\title{Vpc Configuration environment list}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 4.
}
\usage{
vpc_configuration_env
}
\description{
Contains default vpc configurations.
}
\keyword{internal}
