% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{write_bin}
\alias{write_bin}
\title{Write large raw connections in chunks}
\usage{
write_bin(obj, filename, chunk_size = 2^31 - 2)
}
\arguments{
\item{obj}{(raw): raw connection or raw vector}

\item{filename}{(str):}

\item{chunk_size}{(int):}
}
\description{
Write large raw connections in chunks
}
\keyword{internal}
